@@ -60,33 +64,43 @@
     "@storybook/addon-onboarding": "^10.2.1",
     "@storybook/addon-vitest": "^10.2.1",
     "@storybook/react-vite": "^10.2.1",
+    "@tailwindcss/postcss": "^4.1.18",
+    "@tailwindcss/typography": "^0.5.19",
+    "@tailwindcss/vite": "^4.1.18",
+    "@tanstack/react-virtual": "^3.13.18",
+    "@testing-library/jest-dom": "^6.9.1",
+    "@testing-library/react": "^16.3.2",
     "@types/express": "^5.0.6",
-    "@types/node": "^22.10.0",
-    "@types/react": "^18.3.17",
-    "@types/react-dom": "^18.3.5",
+    "@types/node": "^25.1.0",
+    "@types/react": "^19.2.10",
+    "@types/react-dom": "^19.2.3",
     "@types/supertest": "^6.0.3",
     "@types/ws": "^8.18.1",
-    "@vitejs/plugin-react": "^4.3.4",
-    "@vitest/browser": "^3.2.4",
+    "@vitejs/plugin-react": "^5.1.2",
+    "@vitest/browser": "^4.0.18",
+    "@vitest/browser-playwright": "^4.0.18",
     "@xstate/react": "^6.0.0",
-    "autoprefixer": "^10.4.23",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "esbuild": "^0.24.0",
+    "concurrently": "^9.2.1",
+    "esbuild": "^0.27.2",
+    "jsdom": "^27.4.0",
     "lucide-react": "^0.563.0",
     "playwright": "^1.58.0",
     "postcss": "^8.5.6",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-router-dom": "^6.30.0",
+    "react": "^19.2.4",
+    "react-dom": "^19.2.4",
+    "react-markdown": "^10.1.0",
+    "react-router": "^7.13.0",
+    "remark-gfm": "^4.0.1",
     "storybook": "^10.2.1",
     "supertest": "^7.2.2",
     "tailwind-merge": "^3.4.0",
-    "tailwindcss": "^3.4.19",
-    "tailwindcss-animate": "^1.0.7",
+    "tailwindcss": "^4.1.18",
+    "tw-animate-css": "^1.4.0",
     "typescript": "^5.7.0",
-    "vite": "^6.0.5",
-    "vitest": "^3.2.4",
+    "vite": "^7.3.1",
+    "vitest": "^4.0.18",
     "xstate": "^5.26.0"
   },
   "engines": {
