@@ -1,9 +1,28 @@
-const { execSync } = require('child_process');
-const fs = require('fs');
-const path = require('path');
-const os = require('os');
+import process from 'node:process';
+import { execSync } from 'node:child_process';
+import fs from 'node:fs';
+import path from 'node:path';
+import os from 'node:os';
+import { fileURLToPath } from 'node:url';
 
-const SCRIPT_PATH = path.join(__dirname, '..', 'bin', 'task-status');
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
+const SCRIPT_PATH = path.join(__dirname, '..', 'bin', 'saga-status');
+
+// Constants for temp file generation
+const RANDOM_STRING_SLICE_START = 2;
+const BASE_36 = 36;
+
+// Constants for large count tests
+const LARGE_ACTIVE_FEATURES_COUNT = 12;
+const LARGE_DRAFT_FEATURES_COUNT = 8;
+const DRAFT_FEATURE_ID_OFFSET = 100;
+
+// Constants for performance tests
+const NS_PER_MS = 1_000_000;
+const PERFORMANCE_BUDGET_MS = 100;
+const FILE_PERMISSION_DEFAULT = 0o644;
+const TASK_ID_PAD_LENGTH = 3;
+const PERFORMANCE_TEST_FEATURE_COUNT = 10;
 
 /**
  * Helper to run the script with given args and environment
@@ -11,7 +30,7 @@ const SCRIPT_PATH = path.join(__dirname, '..', 'bin', 'task-status');
 function runScript(args = [], env = {}, cwd = undefined) {
   const fullEnv = { ...process.env, ...env };
   // Remove keys that are explicitly set to undefined
-  Object.keys(fullEnv).forEach(key => {
+  Object.keys(fullEnv).forEach((key) => {
     if (fullEnv[key] === undefined) {
       delete fullEnv[key];
     }
@@ -41,8 +60,11 @@ function runScript(args = [], env = {}, cwd = undefined) {
  */
 function createTempEnvFile(content) {
   const tmpDir = os.tmpdir();
-  const tmpFile = path.join(tmpDir, `claude-env-test-${Date.now()}-${Math.random().toString(36).slice(2)}.sh`);
-  fs.writeFileSync(tmpFile, content, { mode: 0o644 });
+  const tmpFile = path.join(
+    tmpDir,
+    `claude-env-test-${Date.now()}-${Math.random().toString(BASE_36).slice(RANDOM_STRING_SLICE_START)}.sh`,
+  );
+  fs.writeFileSync(tmpFile, content, { mode: FILE_PERMISSION_DEFAULT });
   return tmpFile;
 }
 
@@ -61,10 +83,10 @@ function cleanupTempFile(filePath) {
  */
 function createMockFeatures(config = {}) {
   const {
-    features = [],           // Array of { id, name, status } objects
-    noFeaturesDir = false,   // Don't create task-system/features/ at all
+    features = [], // Array of { id, name, status } objects
+    noFeaturesDir = false, // Don't create task-system/features/ at all
     emptyFeaturesDir = false, // Create task-system/features/ but leave it empty
-    malformedFeatures = [],  // Feature IDs with malformed feature.md (missing Status line)
+    malformedFeatures = [], // Feature IDs with malformed feature.md (missing Status line)
   } = config;
 
   const tmpDir = fs.mkdtempSync(path.join(os.tmpdir(), 'features-test-'));
@@ -115,7 +137,9 @@ describe('task-status --counts (feature counts)', () => {
           { id: '002', name: 'payments', status: 'In Progress' },
         ],
       });
-      const envFile = createTempEnvFile(`export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`);
+      const envFile = createTempEnvFile(
+        `export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`,
+      );
 
       try {
         const result = runScript(['--counts', '--no-icons'], { CLAUDE_ENV_FILE: envFile });
@@ -136,7 +160,9 @@ describe('task-status --counts (feature counts)', () => {
           { id: '003', name: 'feature-c', status: 'Draft' },
         ],
       });
-      const envFile = createTempEnvFile(`export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`);
+      const envFile = createTempEnvFile(
+        `export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`,
+      );
 
       try {
         const result = runScript(['--counts', '--no-icons'], { CLAUDE_ENV_FILE: envFile });
@@ -156,7 +182,9 @@ describe('task-status --counts (feature counts)', () => {
           { id: '002', name: 'feature-b', status: 'Planned' },
         ],
       });
-      const envFile = createTempEnvFile(`export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`);
+      const envFile = createTempEnvFile(
+        `export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`,
+      );
 
       try {
         const result = runScript(['--counts', '--no-icons'], { CLAUDE_ENV_FILE: envFile });
@@ -178,7 +206,9 @@ describe('task-status --counts (feature counts)', () => {
           { id: '004', name: 'notifications', status: 'In Progress' },
         ],
       });
-      const envFile = createTempEnvFile(`export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`);
+      const envFile = createTempEnvFile(
+        `export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`,
+      );
 
       try {
         const result = runScript(['--counts', '--no-icons'], { CLAUDE_ENV_FILE: envFile });
@@ -199,7 +229,9 @@ describe('task-status --counts (feature counts)', () => {
           { id: '002', name: 'feature-b', status: 'Draft   ' },
         ],
       });
-      const envFile = createTempEnvFile(`export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`);
+      const envFile = createTempEnvFile(
+        `export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`,
+      );
 
       try {
         const result = runScript(['--counts', '--no-icons'], { CLAUDE_ENV_FILE: envFile });
@@ -222,7 +254,9 @@ describe('task-status --counts (feature counts)', () => {
           { id: '004', name: 'feature-d', status: 'PLANNED' },
         ],
       });
-      const envFile = createTempEnvFile(`export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`);
+      const envFile = createTempEnvFile(
+        `export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`,
+      );
 
       try {
         const result = runScript(['--counts', '--no-icons'], { CLAUDE_ENV_FILE: envFile });
@@ -238,12 +272,12 @@ describe('task-status --counts (feature counts)', () => {
 
     test('should skip features with malformed feature.md (missing Status line)', () => {
       const mockDir = createMockFeatures({
-        features: [
-          { id: '001', name: 'good', status: 'Draft' },
-        ],
+        features: [{ id: '001', name: 'good', status: 'Draft' }],
         malformedFeatures: ['002', '003'],
       });
-      const envFile = createTempEnvFile(`export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`);
+      const envFile = createTempEnvFile(
+        `export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`,
+      );
 
       try {
         const result = runScript(['--counts', '--no-icons'], { CLAUDE_ENV_FILE: envFile });
@@ -265,7 +299,9 @@ describe('task-status --counts (feature counts)', () => {
           { id: '003', name: 'weird', status: 'OnHold' },
         ],
       });
-      const envFile = createTempEnvFile(`export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`);
+      const envFile = createTempEnvFile(
+        `export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`,
+      );
 
       try {
         const result = runScript(['--counts', '--no-icons'], { CLAUDE_ENV_FILE: envFile });
@@ -282,7 +318,9 @@ describe('task-status --counts (feature counts)', () => {
   describe('edge cases - empty and missing directories', () => {
     test('should return 0/0 when task-system/features/ does not exist', () => {
       const mockDir = createMockFeatures({ noFeaturesDir: true });
-      const envFile = createTempEnvFile(`export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`);
+      const envFile = createTempEnvFile(
+        `export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`,
+      );
 
       try {
         const result = runScript(['--counts', '--no-icons'], { CLAUDE_ENV_FILE: envFile });
@@ -298,7 +336,9 @@ describe('task-status --counts (feature counts)', () => {
 
     test('should return 0/0 when task-system/features/ is empty', () => {
       const mockDir = createMockFeatures({ emptyFeaturesDir: true });
-      const envFile = createTempEnvFile(`export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`);
+      const envFile = createTempEnvFile(
+        `export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`,
+      );
 
       try {
         const result = runScript(['--counts', '--no-icons'], { CLAUDE_ENV_FILE: envFile });
@@ -315,7 +355,10 @@ describe('task-status --counts (feature counts)', () => {
       const envFile = createTempEnvFile('export SAGA_TASK_CONTEXT="main"');
 
       try {
-        const result = runScript(['--counts', '--no-icons'], { CLAUDE_ENV_FILE: envFile, SAGA_PROJECT_DIR: undefined });
+        const result = runScript(['--counts', '--no-icons'], {
+          CLAUDE_ENV_FILE: envFile,
+          SAGA_PROJECT_DIR: undefined,
+        });
         expect(result.exitCode).toBe(0);
         // Should output zeros or graceful fallback
         expect(result.stdout).toMatch(/A:0/);
@@ -333,7 +376,9 @@ describe('task-status --counts (feature counts)', () => {
       fs.mkdirSync(path.join(featuresDir, '001-empty'), { recursive: true });
       fs.mkdirSync(path.join(featuresDir, '002-also-empty'), { recursive: true });
 
-      const envFile = createTempEnvFile(`export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${tmpDir}"`);
+      const envFile = createTempEnvFile(
+        `export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${tmpDir}"`,
+      );
 
       try {
         const result = runScript(['--counts', '--no-icons'], { CLAUDE_ENV_FILE: envFile });
@@ -357,14 +402,16 @@ describe('task-status --counts (feature counts)', () => {
           { id: '003', name: 'analytics', status: 'Draft' },
         ],
       });
-      const envFile = createTempEnvFile(`export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`);
+      const envFile = createTempEnvFile(
+        `export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`,
+      );
 
       try {
         const result = runScript(['--counts'], { CLAUDE_ENV_FILE: envFile });
         expect(result.exitCode).toBe(0);
         // Should use Unicode icons: ⭐ for active, 📝 for draft
-        expect(result.stdout).toMatch(/⭐ 1/);   // 1 active
-        expect(result.stdout).toMatch(/📝 2/);   // 2 draft
+        expect(result.stdout).toMatch(/⭐ 1/); // 1 active
+        expect(result.stdout).toMatch(/📝 2/); // 2 draft
       } finally {
         cleanupTempFile(envFile);
         cleanupMockFeatures(mockDir);
@@ -379,7 +426,9 @@ describe('task-status --counts (feature counts)', () => {
           { id: '003', name: 'analytics', status: 'Draft' },
         ],
       });
-      const envFile = createTempEnvFile(`export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`);
+      const envFile = createTempEnvFile(
+        `export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`,
+      );
 
       try {
         const result = runScript(['--counts', '--no-icons'], { CLAUDE_ENV_FILE: envFile });
@@ -395,7 +444,9 @@ describe('task-status --counts (feature counts)', () => {
 
     test('should format all zeros gracefully when no features exist', () => {
       const mockDir = createMockFeatures({ emptyFeaturesDir: true });
-      const envFile = createTempEnvFile(`export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`);
+      const envFile = createTempEnvFile(
+        `export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`,
+      );
 
       try {
         // Unicode version
@@ -416,13 +467,13 @@ describe('task-status --counts (feature counts)', () => {
     });
 
     test('should handle large feature counts correctly', () => {
-      const manyActive = Array.from({ length: 12 }, (_, i) => ({
-        id: String(i + 1).padStart(3, '0'),
+      const manyActive = Array.from({ length: LARGE_ACTIVE_FEATURES_COUNT }, (_, i) => ({
+        id: String(i + 1).padStart(TASK_ID_PAD_LENGTH, '0'),
         name: `active-${i}`,
         status: 'In Progress',
       }));
-      const manyDraft = Array.from({ length: 8 }, (_, i) => ({
-        id: String(i + 100).padStart(3, '0'),
+      const manyDraft = Array.from({ length: LARGE_DRAFT_FEATURES_COUNT }, (_, i) => ({
+        id: String(i + DRAFT_FEATURE_ID_OFFSET).padStart(TASK_ID_PAD_LENGTH, '0'),
         name: `draft-${i}`,
         status: 'Draft',
       }));
@@ -430,7 +481,9 @@ describe('task-status --counts (feature counts)', () => {
       const mockDir = createMockFeatures({
         features: [...manyActive, ...manyDraft],
       });
-      const envFile = createTempEnvFile(`export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`);
+      const envFile = createTempEnvFile(
+        `export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`,
+      );
 
       try {
         const result = runScript(['--counts', '--no-icons'], { CLAUDE_ENV_FILE: envFile });
@@ -457,18 +510,20 @@ describe('task-status --counts (feature counts)', () => {
       const tasksDir = path.join(mockDir, 'task-system', 'tasks');
       fs.mkdirSync(tasksDir, { recursive: true });
 
-      const envFile = createTempEnvFile(`export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`);
+      const envFile = createTempEnvFile(
+        `export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`,
+      );
 
       try {
         const result = runScript(['--counts', '--no-icons'], { CLAUDE_ENV_FILE: envFile });
         expect(result.exitCode).toBe(0);
 
         // Check that both task and feature counts are present
-        expect(result.stdout).toMatch(/I:\d/);  // Task in-progress
-        expect(result.stdout).toMatch(/P:\d/);  // Task pending
-        expect(result.stdout).toMatch(/R:\d/);  // Task remote
-        expect(result.stdout).toMatch(/A:1/);   // Feature active
-        expect(result.stdout).toMatch(/D:1/);   // Feature draft
+        expect(result.stdout).toMatch(/I:\d/); // Task in-progress
+        expect(result.stdout).toMatch(/P:\d/); // Task pending
+        expect(result.stdout).toMatch(/R:\d/); // Task remote
+        expect(result.stdout).toMatch(/A:1/); // Feature active
+        expect(result.stdout).toMatch(/D:1/); // Feature draft
 
         // Feature counts should appear after task counts
         const stdout = result.stdout;
@@ -488,7 +543,9 @@ describe('task-status --counts (feature counts)', () => {
           { id: '002', name: 'payments', status: 'Draft' },
         ],
       });
-      const envFile = createTempEnvFile(`export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`);
+      const envFile = createTempEnvFile(
+        `export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`,
+      );
 
       try {
         const result = runScript(['--counts', '--no-icons'], { CLAUDE_ENV_FILE: envFile });
@@ -515,7 +572,9 @@ describe('task-status --counts (feature counts)', () => {
           { id: '003', name: 'analytics', status: 'Planned' },
         ],
       });
-      const envFile = createTempEnvFile(`export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`);
+      const envFile = createTempEnvFile(
+        `export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`,
+      );
 
       try {
         const result = runScript(['--counts', '--no-icons'], { CLAUDE_ENV_FILE: envFile });
@@ -539,7 +598,9 @@ describe('task-status --counts (feature counts)', () => {
           { id: '005', name: 'reporting', status: 'Planned' },
         ],
       });
-      const envFile = createTempEnvFile(`export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`);
+      const envFile = createTempEnvFile(
+        `export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`,
+      );
 
       try {
         const result = runScript(['--counts', '--no-icons'], { CLAUDE_ENV_FILE: envFile });
@@ -555,12 +616,12 @@ describe('task-status --counts (feature counts)', () => {
 
     test('script handles missing or malformed feature.md files without crashing', () => {
       const mockDir = createMockFeatures({
-        features: [
-          { id: '001', name: 'good', status: 'Draft' },
-        ],
+        features: [{ id: '001', name: 'good', status: 'Draft' }],
         malformedFeatures: ['002', '003'],
       });
-      const envFile = createTempEnvFile(`export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`);
+      const envFile = createTempEnvFile(
+        `export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`,
+      );
 
       try {
         const result = runScript(['--counts', '--no-icons'], { CLAUDE_ENV_FILE: envFile });
@@ -580,13 +641,15 @@ describe('task-status --counts (feature counts)', () => {
           { id: '002', name: 'payments', status: 'Draft' },
         ],
       });
-      const envFile = createTempEnvFile(`export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`);
+      const envFile = createTempEnvFile(
+        `export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`,
+      );
 
       try {
         // Default (icons)
         let result = runScript(['--counts'], { CLAUDE_ENV_FILE: envFile });
-        expect(result.stdout).toMatch(/⭐/);  // Active icon
-        expect(result.stdout).toMatch(/📝/);  // Draft icon
+        expect(result.stdout).toMatch(/⭐/); // Active icon
+        expect(result.stdout).toMatch(/📝/); // Draft icon
 
         // With --no-icons (ASCII)
         result = runScript(['--counts', '--no-icons'], { CLAUDE_ENV_FILE: envFile });
@@ -601,16 +664,16 @@ describe('task-status --counts (feature counts)', () => {
 
     test('feature counts appear after task counts in the output, properly separated', () => {
       const mockDir = createMockFeatures({
-        features: [
-          { id: '001', name: 'auth', status: 'In Progress' },
-        ],
+        features: [{ id: '001', name: 'auth', status: 'In Progress' }],
       });
 
       // Create task structure
       const tasksDir = path.join(mockDir, 'task-system', 'tasks');
       fs.mkdirSync(tasksDir, { recursive: true });
 
-      const envFile = createTempEnvFile(`export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`);
+      const envFile = createTempEnvFile(
+        `export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`,
+      );
 
       try {
         const result = runScript(['--counts', '--no-icons'], { CLAUDE_ENV_FILE: envFile });
@@ -627,7 +690,9 @@ describe('task-status --counts (feature counts)', () => {
         const featureMatch = result.stdout.match(featurePattern);
 
         if (taskMatch && featureMatch) {
-          expect(result.stdout.indexOf(featureMatch[0])).toBeGreaterThan(result.stdout.indexOf(taskMatch[0]));
+          expect(result.stdout.indexOf(featureMatch[0])).toBeGreaterThan(
+            result.stdout.indexOf(taskMatch[0]),
+          );
         }
       } finally {
         cleanupTempFile(envFile);
@@ -638,23 +703,25 @@ describe('task-status --counts (feature counts)', () => {
 
   describe('performance requirements', () => {
     test('should complete within 100ms performance budget with features', () => {
-      const features = Array.from({ length: 10 }, (_, i) => ({
-        id: String(i + 1).padStart(3, '0'),
+      const features = Array.from({ length: PERFORMANCE_TEST_FEATURE_COUNT }, (_, i) => ({
+        id: String(i + 1).padStart(TASK_ID_PAD_LENGTH, '0'),
         name: `feature-${i}`,
         status: i % 2 === 0 ? 'In Progress' : 'Draft',
       }));
 
       const mockDir = createMockFeatures({ features });
-      const envFile = createTempEnvFile(`export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`);
+      const envFile = createTempEnvFile(
+        `export SAGA_TASK_CONTEXT="main"\nexport SAGA_PROJECT_DIR="${mockDir}"`,
+      );
 
       try {
         const start = process.hrtime.bigint();
         const result = runScript(['--counts'], { CLAUDE_ENV_FILE: envFile });
         const end = process.hrtime.bigint();
-        const durationMs = Number(end - start) / 1_000_000;
+        const durationMs = Number(end - start) / NS_PER_MS;
 
         expect(result.exitCode).toBe(0);
-        expect(durationMs).toBeLessThan(100);
+        expect(durationMs).toBeLessThan(PERFORMANCE_BUDGET_MS);
       } finally {
         cleanupTempFile(envFile);
         cleanupMockFeatures(mockDir);
