@@ -1,9 +1,12 @@
-const { execSync, spawn } = require('child_process');
-const fs = require('fs');
-const path = require('path');
-const os = require('os');
+import process from 'node:process';
+import { execSync } from 'node:child_process';
+import fs from 'node:fs';
+import path from 'node:path';
+import os from 'node:os';
+import { fileURLToPath } from 'node:url';
 
-const SCRIPT_PATH = path.join(__dirname, '..', 'bin', 'task-status');
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
+const SCRIPT_PATH = path.join(__dirname, '..', 'bin', 'saga-status');
 
 /**
  * Helper to run the script with given args and environment
@@ -51,7 +54,7 @@ describe('task-status script', () => {
       const result = runScript(['--help']);
       expect(result.exitCode).toBe(0);
       expect(result.stdout).toContain('Usage:');
-      expect(result.stdout).toContain('task-status');
+      expect(result.stdout).toContain('saga-status');
     });
 
     test('should document --no-icons flag', () => {
@@ -96,7 +99,7 @@ describe('task-status script', () => {
           const result = runScript(['--origin'], { CLAUDE_ENV_FILE: envFile });
           expect(result.exitCode).toBe(0);
           // Should contain unicode icon for main repo
-          expect(result.stdout).toMatch(/[🌿🏠]/); // Home or branch icon
+          expect(result.stdout).toMatch(/[🌿🏠]/u); // Home or branch icon
         } finally {
           cleanupTempFile(envFile);
         }
@@ -105,7 +108,9 @@ describe('task-status script', () => {
 
     describe('section selector flags', () => {
       test('--origin should show only origin section', () => {
-        const envFile = createTempEnvFile('export SAGA_TASK_CONTEXT="main"\nexport CURRENT_TASK_ID="042"');
+        const envFile = createTempEnvFile(
+          'export SAGA_TASK_CONTEXT="main"\nexport CURRENT_TASK_ID="042"',
+        );
         try {
           const result = runScript(['--origin'], { CLAUDE_ENV_FILE: envFile });
           expect(result.exitCode).toBe(0);
@@ -117,7 +122,9 @@ describe('task-status script', () => {
       });
 
       test('--task should show only task section', () => {
-        const envFile = createTempEnvFile('export SAGA_TASK_CONTEXT="worktree"\nexport CURRENT_TASK_ID="042"');
+        const envFile = createTempEnvFile(
+          'export SAGA_TASK_CONTEXT="worktree"\nexport CURRENT_TASK_ID="042"',
+        );
         try {
           const result = runScript(['--task'], { CLAUDE_ENV_FILE: envFile });
           expect(result.exitCode).toBe(0);
@@ -138,7 +145,9 @@ describe('task-status script', () => {
       });
 
       test('multiple section flags can be combined', () => {
-        const envFile = createTempEnvFile('export SAGA_TASK_CONTEXT="worktree"\nexport CURRENT_TASK_ID="042"');
+        const envFile = createTempEnvFile(
+          'export SAGA_TASK_CONTEXT="worktree"\nexport CURRENT_TASK_ID="042"',
+        );
         try {
           const result = runScript(['--origin', '--task'], { CLAUDE_ENV_FILE: envFile });
           expect(result.exitCode).toBe(0);
@@ -148,7 +157,9 @@ describe('task-status script', () => {
       });
 
       test('no flags should show all sections (default behavior)', () => {
-        const envFile = createTempEnvFile('export SAGA_TASK_CONTEXT="worktree"\nexport CURRENT_TASK_ID="042"');
+        const envFile = createTempEnvFile(
+          'export SAGA_TASK_CONTEXT="worktree"\nexport CURRENT_TASK_ID="042"',
+        );
         try {
           const result = runScript([], { CLAUDE_ENV_FILE: envFile });
           expect(result.exitCode).toBe(0);
@@ -176,7 +187,9 @@ describe('task-status script', () => {
 
   describe('$CLAUDE_ENV_FILE sourcing', () => {
     test('should source environment file when CLAUDE_ENV_FILE is set', () => {
-      const envFile = createTempEnvFile('export SAGA_TASK_CONTEXT="worktree"\nexport CURRENT_TASK_ID="042"');
+      const envFile = createTempEnvFile(
+        'export SAGA_TASK_CONTEXT="worktree"\nexport CURRENT_TASK_ID="042"',
+      );
       try {
         const result = runScript(['--task'], { CLAUDE_ENV_FILE: envFile });
         expect(result.exitCode).toBe(0);
@@ -230,7 +243,7 @@ describe('task-status script', () => {
           const result = runScript(['--origin'], { CLAUDE_ENV_FILE: envFile });
           expect(result.exitCode).toBe(0);
           // Unicode: home icon 🏠
-          expect(result.stdout).toMatch(/[🏠]/);
+          expect(result.stdout).toMatch(/[🏠]/u);
         } finally {
           cleanupTempFile(envFile);
         }
@@ -266,7 +279,7 @@ describe('task-status script', () => {
           const result = runScript(['--origin'], { CLAUDE_ENV_FILE: envFile });
           expect(result.exitCode).toBe(0);
           // Unicode: branch icon 🌿
-          expect(result.stdout).toMatch(/[🌿]/);
+          expect(result.stdout).toMatch(/[🌿]/u);
         } finally {
           cleanupTempFile(envFile);
         }
