@@ -1,4 +1,11 @@
+import process from 'node:process';
 import { defineConfig, devices } from '@playwright/test';
+
+// Timeout configuration constants
+const MS_PER_SECOND = 1000;
+const WEB_SERVER_TIMEOUT_SECONDS = 120;
+const TEST_TIMEOUT_SECONDS = 30;
+const EXPECT_TIMEOUT_SECONDS = 5;
 
 /**
  * Playwright configuration for dashboard integration tests.
@@ -14,20 +21,20 @@ export default defineConfig({
   fullyParallel: true,
 
   // Fail the build on CI if you accidentally left test.only in the source code
-  forbidOnly: !!process.env.CI,
+  forbidOnly: Boolean(process.env.CI),
 
   // Retry on CI only
   retries: process.env.CI ? 2 : 0,
 
-  // Limit parallel workers on CI for stability
-  workers: process.env.CI ? 1 : undefined,
+  // Stop on first failure to fail fast
+  maxFailures: 1,
 
   // Reporter configuration
   reporter: process.env.CI ? 'github' : 'html',
 
   // Shared settings for all projects
   use: {
-    // Base URL for navigation
+    // Base URL for navigation (Playwright API uses baseURL)
     baseURL: 'http://localhost:5173',
 
     // Collect trace on first retry
@@ -47,16 +54,16 @@ export default defineConfig({
 
   // Web server configuration - starts Vite dev server before tests
   webServer: {
-    command: 'pnpm dev:client',
+    command: 'vite --config src/client/vite.config.ts',
     cwd: '../..',
     url: 'http://localhost:5173',
     reuseExistingServer: !process.env.CI,
-    timeout: 120 * 1000,
+    timeout: WEB_SERVER_TIMEOUT_SECONDS * MS_PER_SECOND,
   },
 
   // Test timeouts
-  timeout: 30 * 1000,
+  timeout: TEST_TIMEOUT_SECONDS * MS_PER_SECOND,
   expect: {
-    timeout: 5 * 1000,
+    timeout: EXPECT_TIMEOUT_SECONDS * MS_PER_SECOND,
   },
 });
