@@ -1,14 +1,26 @@
-import { Outlet } from 'react-router-dom';
-import { Breadcrumb } from './Breadcrumb';
+import { useEffect, useRef } from 'react';
+import { Outlet } from 'react-router';
 import { Toaster } from '@/components/ui/toaster';
+import { useDashboard } from '@/context/dashboard-context';
 import { useDashboardToasts } from '@/hooks/use-dashboard-toasts';
+import { Breadcrumb } from './Breadcrumb.tsx';
 
 export function Layout() {
   // Set up toast notifications for dashboard state changes
   useDashboardToasts();
 
+  // Auto-connect to WebSocket on mount for real-time updates
+  const { connect, isConnected } = useDashboard();
+  const hasConnected = useRef(false);
+  useEffect(() => {
+    if (!hasConnected.current) {
+      hasConnected.current = true;
+      connect();
+    }
+  }, [connect]);
+
   return (
-    <div className="min-h-screen bg-bg">
+    <div className="min-h-screen bg-bg" data-ws-connected={isConnected}>
       <header className="border-b border-border-muted bg-bg-dark">
         <div className="container mx-auto px-4 py-4">
           <div className="flex items-center justify-between">
@@ -28,5 +40,3 @@ export function Layout() {
     </div>
   );
 }
-
-export default Layout;
