@@ -1,13 +1,18 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
-import { AppRouter } from './router';
-import { DashboardProvider } from './context/DashboardContext';
+import { DashboardProvider } from './context/dashboard-context.tsx';
+import { AppRouter } from './router.tsx';
 import './index.css';
 
-createRoot(document.getElementById('root')!).render(
+const rootElement = document.getElementById('root');
+if (!rootElement) {
+  throw new Error('Root element not found');
+}
+
+createRoot(rootElement).render(
   <StrictMode>
     <DashboardProvider>
       <AppRouter />
     </DashboardProvider>
-  </StrictMode>
+  </StrictMode>,
 );
