@@ -1,20 +1,27 @@
+import { AlertCircle, CheckCircle, ChevronDown, ChevronRight, Circle } from 'lucide-react';
 import { useEffect, useState } from 'react';
-import { Link, useParams } from 'react-router-dom';
-import { useDashboard } from '@/context/DashboardContext';
-import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
+import ReactMarkdown from 'react-markdown';
+import { Link, useParams } from 'react-router';
+import remarkGfm from 'remark-gfm';
 import { Badge } from '@/components/ui/badge';
+import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
+import { Collapsible, CollapsibleContent, CollapsibleTrigger } from '@/components/ui/collapsible';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
-import {
-  Collapsible,
-  CollapsibleContent,
-  CollapsibleTrigger,
-} from '@/components/ui/collapsible';
+import { useDashboard } from '@/context/dashboard-context';
 import { showApiErrorToast } from '@/lib/toast-utils';
-import type { StoryDetail as StoryDetailType, StoryStatus, TaskStatus, JournalEntry, JournalEntryType } from '@/types/dashboard';
-import { ChevronDown, ChevronRight, CheckCircle, Circle, AlertCircle } from 'lucide-react';
+import type {
+  JournalEntry,
+  JournalEntryType,
+  StoryDetail as StoryDetailType,
+  StoryStatus,
+  TaskStatus,
+} from '@/types/dashboard';
+
+/** HTTP 404 Not Found status code */
+const HTTP_NOT_FOUND = 404;
 
 /** Skeleton loading component for the story header */
-export function HeaderSkeleton() {
+function HeaderSkeleton() {
   return (
     <div className="animate-pulse space-y-4" data-testid="story-header-skeleton">
       <div className="h-8 w-64 bg-bg-light rounded" />
@@ -27,7 +34,7 @@ export function HeaderSkeleton() {
 }
 
 /** Skeleton loading component for content sections */
-export function ContentSkeleton() {
+function ContentSkeleton() {
   return (
     <div className="animate-pulse space-y-4" data-testid="story-content-skeleton">
       <div className="h-6 w-48 bg-bg-light rounded" />
@@ -41,17 +48,17 @@ export function ContentSkeleton() {
 }
 
 /** Status badge with appropriate color based on story status */
-export function StatusBadge({ status }: { status: StoryStatus }) {
+function StatusBadge({ status }: { status: StoryStatus }) {
   const variants: Record<StoryStatus, string> = {
     ready: 'bg-text-muted/20 text-text-muted',
-    in_progress: 'bg-primary/20 text-primary',
+    inProgress: 'bg-primary/20 text-primary',
     blocked: 'bg-danger/20 text-danger',
     completed: 'bg-success/20 text-success',
   };
 
   const labels: Record<StoryStatus, string> = {
     ready: 'Ready',
-    in_progress: 'In Progress',
+    inProgress: 'In Progress',
     blocked: 'Blocked',
     completed: 'Completed',
   };
@@ -60,39 +67,67 @@ export function StatusBadge({ status }: { status: StoryStatus }) {
 }
 
 /** Task status icon (visual only, not interactive) */
-export function TaskStatusIcon({ status }: { status: TaskStatus }) {
+function TaskStatusIcon({ status }: { status: TaskStatus }) {
   const iconProps = { className: 'w-5 h-5 pointer-events-none cursor-default' };
 
   switch (status) {
     case 'completed':
-      return <CheckCircle {...iconProps} className={`${iconProps.className} text-success`} />;
-    case 'in_progress':
-      return <Circle {...iconProps} className={`${iconProps.className} text-primary fill-primary/20`} />;
-    case 'pending':
+      return (
+        <CheckCircle
+          {...iconProps}
+          className={`${iconProps.className} text-success`}
+          data-testid="icon-check-circle"
+        />
+      );
+    case 'inProgress':
+      return (
+        <Circle
+          {...iconProps}
+          className={`${iconProps.className} text-primary fill-primary/20`}
+          data-testid="icon-circle-in-progress"
+        />
+      );
     default:
-      return <Circle {...iconProps} className={`${iconProps.className} text-text-muted`} />;
+      return (
+        <Circle
+          {...iconProps}
+          className={`${iconProps.className} text-text-muted`}
+          data-testid="icon-circle-pending"
+        />
+      );
   }
 }
 
+/** Task status badge styles */
+const taskBadgeStyles: Record<TaskStatus, string> = {
+  pending: 'bg-text-muted/20 text-text-muted',
+  inProgress: 'bg-primary/20 text-primary',
+  completed: 'bg-success/20 text-success',
+};
+
+/** Task status labels */
+const taskStatusLabels: Record<TaskStatus, string> = {
+  pending: 'pending',
+  inProgress: 'in progress',
+  completed: 'completed',
+};
+
+/** Get badge class based on task status */
+function getTaskBadgeClass(status: TaskStatus): string {
+  return taskBadgeStyles[status] || taskBadgeStyles.pending;
+}
+
 /** Single task item display */
-export function TaskItem({ task }: { task: StoryDetailType['tasks'][0] }) {
+function TaskItem({ task }: { task: StoryDetailType['tasks'][0] }) {
+  const badgeClass = getTaskBadgeClass(task.status);
+
   return (
     <div className="flex items-center gap-3 py-2 px-3 rounded-md hover:bg-bg-light/50">
       <TaskStatusIcon status={task.status} />
       <span className={task.status === 'completed' ? 'text-text-muted line-through' : 'text-text'}>
         {task.title}
       </span>
-      <Badge
-        className={`ml-auto text-xs ${
-          task.status === 'completed'
-            ? 'bg-success/20 text-success'
-            : task.status === 'in_progress'
-            ? 'bg-primary/20 text-primary'
-            : 'bg-text-muted/20 text-text-muted'
-        }`}
-      >
-        {task.status.replace('_', ' ')}
-      </Badge>
+      <Badge className={`ml-auto text-xs ${badgeClass}`}>{taskStatusLabels[task.status]}</Badge>
     </div>
   );
 }
@@ -104,41 +139,48 @@ function getEntryTypeStyle(type: JournalEntryType): { bg: string; text: string; 
       return { bg: 'bg-danger/10', text: 'text-danger', border: 'border-danger/30' };
     case 'resolution':
       return { bg: 'bg-success/10', text: 'text-success', border: 'border-success/30' };
-    case 'session':
     default:
       return { bg: 'bg-bg-light', text: 'text-text', border: 'border-border-muted' };
   }
 }
 
 /** Single journal entry with collapsible content */
-export function JournalEntryItem({ entry, defaultOpen = false }: { entry: JournalEntry; defaultOpen?: boolean }) {
+function JournalEntryItem({
+  entry,
+  defaultOpen = false,
+}: {
+  entry: JournalEntry;
+  defaultOpen?: boolean;
+}) {
   const [isOpen, setIsOpen] = useState(defaultOpen);
   const style = getEntryTypeStyle(entry.type);
 
   return (
     <Collapsible open={isOpen} onOpenChange={setIsOpen}>
       <Card className={`${style.bg} border ${style.border}`}>
-        <CollapsibleTrigger asChild>
+        <CollapsibleTrigger asChild={true}>
           <CardHeader className="cursor-pointer hover:bg-bg-light/50 transition-colors py-3">
             <div className="flex items-center gap-3">
               {isOpen ? (
-                <ChevronDown className="w-4 h-4 text-text-muted" />
+                <ChevronDown className="w-4 h-4 text-text-muted" data-testid="icon-chevron-down" />
               ) : (
-                <ChevronRight className="w-4 h-4 text-text-muted" />
+                <ChevronRight
+                  className="w-4 h-4 text-text-muted"
+                  data-testid="icon-chevron-right"
+                />
               )}
               <Badge className={`${style.bg} ${style.text} border ${style.border}`}>
                 {entry.type}
               </Badge>
-              <CardTitle className={`text-sm font-medium ${style.text}`}>
-                {entry.title}
-              </CardTitle>
+              <CardTitle className={`text-sm font-medium ${style.text}`}>{entry.title}</CardTitle>
               {entry.type === 'blocker' && (
-                <AlertCircle className="w-4 h-4 text-danger ml-auto" />
+                <AlertCircle
+                  className="w-4 h-4 text-danger ml-auto"
+                  data-testid="icon-alert-circle"
+                />
               )}
               {entry.timestamp && (
-                <span className="text-xs text-text-muted ml-auto">
-                  {entry.timestamp}
-                </span>
+                <span className="text-xs text-text-muted ml-auto">{entry.timestamp}</span>
               )}
             </div>
           </CardHeader>
@@ -157,47 +199,264 @@ export function JournalEntryItem({ entry, defaultOpen = false }: { entry: Journa
   );
 }
 
-export function StoryDetail() {
-  const { epicSlug, storySlug } = useParams<{
-    epicSlug: string;
-    storySlug: string;
-  }>();
+/** Calculate task completion progress */
+function calculateTaskProgress(story: StoryDetailType | null) {
+  if (!story) {
+    return { completed: 0, total: 0 };
+  }
+  return {
+    completed: story.tasks.filter((t) => t.status === 'completed').length,
+    total: story.tasks.length,
+  };
+}
+
+/** Group journal entries by type */
+function groupJournalEntries(journal: JournalEntry[]) {
+  return {
+    blockers: journal.filter((e) => e.type === 'blocker'),
+    sessions: journal.filter((e) => e.type === 'session'),
+    resolutions: journal.filter((e) => e.type === 'resolution'),
+  };
+}
+
+/** Story not found state */
+function StoryNotFoundState({ epicSlug, storySlug }: { epicSlug: string; storySlug: string }) {
+  return (
+    <div className="text-center py-12">
+      <h1 className="text-2xl font-bold text-text mb-2">Story not found</h1>
+      <p className="text-text-muted mb-4">
+        The story &quot;{storySlug}&quot; does not exist in epic &quot;{epicSlug}&quot;.
+      </p>
+      <Link to={`/epic/${epicSlug}`} className="text-primary hover:underline">
+        ← Back to epic
+      </Link>
+    </div>
+  );
+}
+
+/** Story error state */
+function StoryErrorState({ epicSlug, error }: { epicSlug: string; error: string }) {
+  return (
+    <div className="text-center py-12">
+      <h1 className="text-2xl font-bold text-danger mb-2">Error</h1>
+      <p className="text-text-muted mb-4">{error}</p>
+      <Link to={`/epic/${epicSlug}`} className="text-primary hover:underline">
+        ← Back to epic
+      </Link>
+    </div>
+  );
+}
+
+/** Story loading state */
+function StoryLoadingState() {
+  return (
+    <div className="space-y-6">
+      <HeaderSkeleton />
+      <ContentSkeleton />
+    </div>
+  );
+}
+
+/** Story header with breadcrumb, title, and status */
+function StoryHeader({
+  story,
+  epicSlug,
+  storySlug,
+}: {
+  story: StoryDetailType;
+  epicSlug: string;
+  storySlug: string;
+}) {
+  const taskProgress = calculateTaskProgress(story);
+
+  return (
+    <div className="space-y-4">
+      <div className="flex items-center gap-2 text-sm text-text-muted">
+        <Link to={`/epic/${epicSlug}`} className="hover:text-primary">
+          {epicSlug}
+        </Link>
+        <span>/</span>
+        <span className="text-text">{storySlug}</span>
+      </div>
+      <h1 className="text-2xl font-bold text-text">{story.title}</h1>
+      <div className="flex items-center gap-4">
+        <StatusBadge status={story.status} />
+        <span className="text-sm text-text-muted">
+          {taskProgress.completed}/{taskProgress.total} tasks completed
+        </span>
+      </div>
+    </div>
+  );
+}
+
+/** Tasks tab content */
+function TasksTabContent({ tasks }: { tasks: StoryDetailType['tasks'] }) {
+  return (
+    <Card>
+      <CardHeader>
+        <CardTitle className="text-lg">Tasks</CardTitle>
+      </CardHeader>
+      <CardContent>
+        {tasks.length === 0 ? (
+          <p className="text-text-muted text-center py-4">No tasks defined for this story.</p>
+        ) : (
+          <div className="divide-y divide-border-muted">
+            {tasks.map((task) => (
+              <TaskItem key={task.id} task={task} />
+            ))}
+          </div>
+        )}
+      </CardContent>
+    </Card>
+  );
+}
+
+/** Story content tab */
+function ContentTabContent({ content }: { content: string | undefined }) {
+  return (
+    <Card>
+      <CardHeader>
+        <CardTitle className="text-lg">Story Content</CardTitle>
+      </CardHeader>
+      <CardContent>
+        {content ? (
+          <div className="prose prose-sm prose-invert max-w-none prose-headings:text-text prose-p:text-text-muted prose-strong:text-text prose-code:text-primary prose-code:bg-bg-dark prose-code:px-1 prose-code:py-0.5 prose-code:rounded prose-code:before:content-none prose-code:after:content-none prose-pre:bg-bg-dark prose-pre:border prose-pre:border-border-muted prose-a:text-primary prose-a:no-underline prose-a:hover:underline prose-li:text-text-muted prose-table:border prose-table:border-border-muted prose-th:bg-bg-dark prose-th:px-3 prose-th:py-2 prose-th:text-text prose-td:px-3 prose-td:py-2 prose-td:border-t prose-td:border-border-muted">
+            <ReactMarkdown remarkPlugins={[remarkGfm]}>{content}</ReactMarkdown>
+          </div>
+        ) : (
+          <p className="text-text-muted text-center py-4">No story content available.</p>
+        )}
+      </CardContent>
+    </Card>
+  );
+}
+
+/** Get color class for journal section type */
+function getJournalSectionColor(type: string): string {
+  if (type === 'Blockers') {
+    return 'text-danger';
+  }
+  if (type === 'Resolutions') {
+    return 'text-success';
+  }
+  return 'text-text';
+}
+
+/** Journal entries section */
+function JournalSection({
+  entries,
+  type,
+  icon: Icon,
+}: {
+  entries: JournalEntry[];
+  type: string;
+  icon: typeof AlertCircle;
+}) {
+  if (entries.length === 0) {
+    return null;
+  }
+
+  const colorClass = getJournalSectionColor(type);
+
+  return (
+    <div className="space-y-3">
+      <h3 className={`text-sm font-semibold ${colorClass} flex items-center gap-2`}>
+        <Icon className="w-4 h-4" />
+        {type} ({entries.length})
+      </h3>
+      {entries.map((entry) => (
+        <JournalEntryItem
+          key={`${type.toLowerCase()}-${entry.title}-${entry.timestamp ?? ''}`}
+          entry={entry}
+          defaultOpen={type === 'Blockers'}
+        />
+      ))}
+    </div>
+  );
+}
+
+/** Journal tab content */
+function JournalTabContent({ journal }: { journal: JournalEntry[] }) {
+  if (journal.length === 0) {
+    return (
+      <Card>
+        <CardContent className="py-8">
+          <p className="text-text-muted text-center">No journal entries yet.</p>
+        </CardContent>
+      </Card>
+    );
+  }
+
+  const { blockers, sessions, resolutions } = groupJournalEntries(journal);
+
+  return (
+    <div className="space-y-4">
+      <JournalSection entries={blockers} type="Blockers" icon={AlertCircle} />
+      <JournalSection entries={resolutions} type="Resolutions" icon={CheckCircle} />
+      <JournalSection entries={sessions} type="Sessions" icon={Circle} />
+    </div>
+  );
+}
+
+/** Result type for fetch response processing */
+type FetchResult = { type: 'notFound' } | { type: 'error' } | { type: 'success'; data: unknown };
+
+/** Process fetch response into a result type */
+async function processFetchResponse(response: Response): Promise<FetchResult> {
+  if (response.status === HTTP_NOT_FOUND) {
+    return { type: 'notFound' };
+  }
+  if (!response.ok) {
+    return { type: 'error' };
+  }
+  return { type: 'success', data: await response.json() };
+}
+
+/** Handle fetch error with toast notification */
+function handleFetchError(url: string, err: unknown, setError: (e: string) => void): void {
+  setError('Failed to load story');
+  showApiErrorToast(url, err instanceof Error ? err.message : 'Unknown error');
+}
+
+/** Custom hook for fetching story data */
+function useStoryFetch(epicSlug: string | undefined, storySlug: string | undefined) {
   const {
     currentStory,
     setCurrentStory,
     clearCurrentStory,
-    isLoading,
     subscribeToStory,
     unsubscribeFromStory,
+    isConnected,
   } = useDashboard();
   const [isFetching, setIsFetching] = useState(true);
   const [notFound, setNotFound] = useState(false);
   const [error, setError] = useState<string | null>(null);
+  const hasSlugs = Boolean(epicSlug && storySlug);
 
+  // Fetch story data on mount
   useEffect(() => {
-    const fetchStory = async () => {
-      if (!epicSlug || !storySlug) return;
+    if (!hasSlugs) {
+      return clearCurrentStory;
+    }
 
+    const fetchStory = async () => {
       setIsFetching(true);
       setNotFound(false);
       setError(null);
 
       try {
         const response = await fetch(`/api/stories/${epicSlug}/${storySlug}`);
-        if (response.status === 404) {
+        const result = await processFetchResponse(response);
+
+        if (result.type === 'notFound') {
           setNotFound(true);
-          return;
-        }
-        if (!response.ok) {
+        } else if (result.type === 'error') {
           setError('Failed to load story');
-          return;
+        } else {
+          setCurrentStory(result.data);
         }
-        const data: StoryDetailType = await response.json();
-        setCurrentStory(data);
       } catch (err) {
-        setError('Failed to load story');
-        const message = err instanceof Error ? err.message : 'Unknown error';
-        showApiErrorToast(`/api/stories/${epicSlug}/${storySlug}`, message);
+        handleFetchError(`/api/stories/${epicSlug}/${storySlug}`, err, setError);
       } finally {
         setIsFetching(false);
       }
@@ -205,212 +464,75 @@ export function StoryDetail() {
 
     fetchStory();
 
-    // Subscribe to real-time story updates
-    if (epicSlug && storySlug) {
-      subscribeToStory(epicSlug, storySlug);
+    return clearCurrentStory;
+  }, [epicSlug, storySlug, hasSlugs, setCurrentStory, clearCurrentStory]);
+
+  // Subscribe to story updates when WebSocket is connected
+  useEffect(() => {
+    if (!(hasSlugs && isConnected)) {
+      return;
     }
 
+    subscribeToStory(epicSlug as string, storySlug as string);
+
     return () => {
-      clearCurrentStory();
-      if (epicSlug && storySlug) {
-        unsubscribeFromStory(epicSlug, storySlug);
-      }
+      unsubscribeFromStory(epicSlug as string, storySlug as string);
     };
-  }, [epicSlug, storySlug, setCurrentStory, clearCurrentStory, subscribeToStory, unsubscribeFromStory]);
+  }, [epicSlug, storySlug, hasSlugs, isConnected, subscribeToStory, unsubscribeFromStory]);
 
-  const loading = isLoading || isFetching;
+  return { currentStory, loading: isFetching, notFound, error };
+}
 
-  // Calculate task progress
-  const taskProgress = currentStory
-    ? {
-        completed: currentStory.tasks.filter((t) => t.status === 'completed').length,
-        total: currentStory.tasks.length,
-      }
-    : { completed: 0, total: 0 };
+function StoryDetail() {
+  const { epicSlug, storySlug } = useParams<{ epicSlug: string; storySlug: string }>();
+  const { currentStory, loading, notFound, error } = useStoryFetch(epicSlug, storySlug);
 
-  // Group journal entries by type
-  const blockerEntries = currentStory?.journal.filter((e) => e.type === 'blocker') || [];
-  const sessionEntries = currentStory?.journal.filter((e) => e.type === 'session') || [];
-  const resolutionEntries = currentStory?.journal.filter((e) => e.type === 'resolution') || [];
-
-  // 404 state
   if (notFound) {
-    return (
-      <div className="text-center py-12">
-        <h1 className="text-2xl font-bold text-text mb-2">Story not found</h1>
-        <p className="text-text-muted mb-4">
-          The story &quot;{storySlug}&quot; does not exist in epic &quot;{epicSlug}&quot;.
-        </p>
-        <Link to={`/epic/${epicSlug}`} className="text-primary hover:underline">
-          ← Back to epic
-        </Link>
-      </div>
-    );
+    return <StoryNotFoundState epicSlug={epicSlug ?? ''} storySlug={storySlug ?? ''} />;
   }
-
-  // Error state
   if (error && !loading) {
-    return (
-      <div className="text-center py-12">
-        <h1 className="text-2xl font-bold text-danger mb-2">Error</h1>
-        <p className="text-text-muted mb-4">{error}</p>
-        <Link to={`/epic/${epicSlug}`} className="text-primary hover:underline">
-          ← Back to epic
-        </Link>
-      </div>
-    );
+    return <StoryErrorState epicSlug={epicSlug ?? ''} error={error} />;
+  }
+  if (loading || !currentStory) {
+    return <StoryLoadingState />;
   }
 
-  // Loading state
-  if (loading || !currentStory) {
-    return (
-      <div className="space-y-6">
-        <HeaderSkeleton />
-        <ContentSkeleton />
-      </div>
-    );
-  }
+  const { blockers } = groupJournalEntries(currentStory.journal);
 
   return (
     <div className="space-y-6">
-      {/* Story header */}
-      <div className="space-y-4">
-        <div className="flex items-center gap-2 text-sm text-text-muted">
-          <Link to={`/epic/${epicSlug}`} className="hover:text-primary">
-            {epicSlug}
-          </Link>
-          <span>/</span>
-          <span className="text-text">{storySlug}</span>
-        </div>
-        <h1 className="text-2xl font-bold text-text">{currentStory.title}</h1>
-        <div className="flex items-center gap-4">
-          <StatusBadge status={currentStory.status} />
-          <span className="text-sm text-text-muted">
-            {taskProgress.completed}/{taskProgress.total} tasks completed
-          </span>
-        </div>
-      </div>
-
-      {/* Tabs for Story Content and Journal */}
+      <StoryHeader story={currentStory} epicSlug={epicSlug ?? ''} storySlug={storySlug ?? ''} />
       <Tabs defaultValue="tasks" className="w-full">
         <TabsList className="mb-4">
           <TabsTrigger value="tasks">Tasks</TabsTrigger>
           <TabsTrigger value="content">Story Content</TabsTrigger>
           <TabsTrigger value="journal">
             Journal
-            {blockerEntries.length > 0 && (
-              <Badge className="ml-2 bg-danger/20 text-danger text-xs">
-                {blockerEntries.length}
-              </Badge>
+            {blockers.length > 0 && (
+              <Badge className="ml-2 bg-danger/20 text-danger text-xs">{blockers.length}</Badge>
             )}
           </TabsTrigger>
         </TabsList>
-
-        {/* Tasks tab */}
         <TabsContent value="tasks" className="space-y-4">
-          <Card>
-            <CardHeader>
-              <CardTitle className="text-lg">Tasks</CardTitle>
-            </CardHeader>
-            <CardContent>
-              {currentStory.tasks.length === 0 ? (
-                <p className="text-text-muted text-center py-4">No tasks defined for this story.</p>
-              ) : (
-                <div className="divide-y divide-border-muted">
-                  {currentStory.tasks.map((task) => (
-                    <TaskItem key={task.id} task={task} />
-                  ))}
-                </div>
-              )}
-            </CardContent>
-          </Card>
+          <TasksTabContent tasks={currentStory.tasks} />
         </TabsContent>
-
-        {/* Story Content tab */}
         <TabsContent value="content" className="space-y-4">
-          <Card>
-            <CardHeader>
-              <CardTitle className="text-lg">Story Content</CardTitle>
-            </CardHeader>
-            <CardContent>
-              {currentStory.content ? (
-                <div className="prose prose-sm prose-invert max-w-none">
-                  <pre className="whitespace-pre-wrap text-sm text-text font-mono bg-bg-dark p-4 rounded-md overflow-x-auto">
-                    {currentStory.content}
-                  </pre>
-                </div>
-              ) : (
-                <p className="text-text-muted text-center py-4">No story content available.</p>
-              )}
-            </CardContent>
-          </Card>
+          <ContentTabContent content={currentStory.content} />
         </TabsContent>
-
-        {/* Journal tab */}
         <TabsContent value="journal" className="space-y-4">
-          {currentStory.journal.length === 0 ? (
-            <Card>
-              <CardContent className="py-8">
-                <p className="text-text-muted text-center">No journal entries yet.</p>
-              </CardContent>
-            </Card>
-          ) : (
-            <div className="space-y-4">
-              {/* Blockers section - highlighted prominently */}
-              {blockerEntries.length > 0 && (
-                <div className="space-y-3">
-                  <h3 className="text-sm font-semibold text-danger flex items-center gap-2">
-                    <AlertCircle className="w-4 h-4" />
-                    Blockers ({blockerEntries.length})
-                  </h3>
-                  {blockerEntries.map((entry, index) => (
-                    <JournalEntryItem
-                      key={`blocker-${index}`}
-                      entry={entry}
-                      defaultOpen={true}
-                    />
-                  ))}
-                </div>
-              )}
-
-              {/* Resolutions section */}
-              {resolutionEntries.length > 0 && (
-                <div className="space-y-3">
-                  <h3 className="text-sm font-semibold text-success flex items-center gap-2">
-                    <CheckCircle className="w-4 h-4" />
-                    Resolutions ({resolutionEntries.length})
-                  </h3>
-                  {resolutionEntries.map((entry, index) => (
-                    <JournalEntryItem
-                      key={`resolution-${index}`}
-                      entry={entry}
-                      defaultOpen={false}
-                    />
-                  ))}
-                </div>
-              )}
-
-              {/* Sessions section */}
-              {sessionEntries.length > 0 && (
-                <div className="space-y-3">
-                  <h3 className="text-sm font-semibold text-text flex items-center gap-2">
-                    Sessions ({sessionEntries.length})
-                  </h3>
-                  {sessionEntries.map((entry, index) => (
-                    <JournalEntryItem
-                      key={`session-${index}`}
-                      entry={entry}
-                      defaultOpen={false}
-                    />
-                  ))}
-                </div>
-              )}
-            </div>
-          )}
+          <JournalTabContent journal={currentStory.journal} />
         </TabsContent>
       </Tabs>
     </div>
   );
 }
 
-export default StoryDetail;
+export {
+  HeaderSkeleton,
+  ContentSkeleton,
+  StatusBadge,
+  TaskStatusIcon,
+  TaskItem,
+  JournalEntryItem,
+  StoryDetail,
+};
