@@ -1,7 +1,7 @@
-import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import { BrowserRouter, Route, Routes } from 'react-router';
 import { Layout } from '@/components/Layout';
-import { EpicList } from '@/pages/EpicList';
 import { EpicDetail } from '@/pages/EpicDetail';
+import { EpicList } from '@/pages/EpicList';
 import { StoryDetail } from '@/pages/StoryDetail';
 
 const routes = [
@@ -37,5 +37,3 @@ export function AppRouter() {
     </BrowserRouter>
   );
 }
-
-export default AppRouter;
