@@ -4,10 +4,10 @@
  */
 
 /** Task status values */
-export type TaskStatus = 'pending' | 'in_progress' | 'completed';
+export type TaskStatus = 'pending' | 'inProgress' | 'completed';
 
 /** Story status values */
-export type StoryStatus = 'ready' | 'in_progress' | 'blocked' | 'completed';
+export type StoryStatus = 'ready' | 'inProgress' | 'blocked' | 'completed';
 
 /** Journal entry types */
 export type JournalEntryType = 'session' | 'blocker' | 'resolution';
@@ -63,4 +63,32 @@ export interface Epic {
   stories: StoryDetail[];
   storyCounts: StoryCounts;
   isArchived?: boolean;
+}
+
+/** Session status values */
+export type SessionStatus = 'running' | 'completed';
+
+/**
+ * Session info from the dashboard API
+ * Matches the backend DetailedSessionInfo type with JSON-serialized dates
+ */
+export interface SessionInfo {
+  /** Unique session name (saga__<epic>__<story>__<pid>) */
+  name: string;
+  /** Epic slug extracted from session name */
+  epicSlug: string;
+  /** Story slug extracted from session name */
+  storySlug: string;
+  /** Current session status */
+  status: SessionStatus;
+  /** Path to the output file */
+  outputFile: string;
+  /** Whether the output file exists and is readable */
+  outputAvailable: boolean;
+  /** Session start time (ISO 8601 string) */
+  startTime: string;
+  /** Session end time (ISO 8601 string), only present for completed sessions */
+  endTime?: string;
+  /** Preview of the last 5 lines of output (max 500 chars) */
+  outputPreview?: string;
 }
