@@ -1,14 +1,25 @@
-import { test, expect } from '@playwright/test';
+import { expect } from '@playwright/test';
+import { test } from '../fixtures.ts';
 import {
-  mockEpicList,
-  mockEpicDetail,
-  mockStoryDetail,
-  mockApiError,
-  mockApiDelay,
-  createMockEpicSummary,
   createMockEpic,
+  createMockEpicSummary,
   createMockStoryDetail,
-} from '../utils/mock-api';
+  mockApiDelay,
+  mockApiError,
+  mockEpicDetail,
+  mockEpicList,
+  mockStoryDetail,
+} from '../utils/mock-api.ts';
+
+// HTTP Status Codes
+const HTTP_OK = 200;
+const HTTP_INTERNAL_SERVER_ERROR = 500;
+
+// Timeouts (ms)
+const LOADING_TIMEOUT_MS = 10_000;
+
+// Delay durations (ms)
+const DELAY_MS = 500;
 
 /**
  * Tests for API mocking infrastructure.
@@ -16,7 +27,7 @@ import {
  */
 test.describe('API Mocking Infrastructure', () => {
   test.describe('Mock Data Factories', () => {
-    test('createMockEpicSummary creates valid EpicSummary', async () => {
+    test('createMockEpicSummary creates valid EpicSummary', () => {
       const epic = createMockEpicSummary();
       expect(epic.slug).toBeDefined();
       expect(epic.title).toBeDefined();
@@ -24,7 +35,7 @@ test.describe('API Mocking Infrastructure', () => {
       expect(epic.storyCounts.total).toBeGreaterThanOrEqual(0);
     });
 
-    test('createMockEpicSummary allows overrides', async () => {
+    test('createMockEpicSummary allows overrides', () => {
       const epic = createMockEpicSummary({
         slug: 'custom-slug',
         title: 'Custom Title',
@@ -35,7 +46,7 @@ test.describe('API Mocking Infrastructure', () => {
       expect(epic.isArchived).toBe(true);
     });
 
-    test('createMockEpic creates valid Epic with stories', async () => {
+    test('createMockEpic creates valid Epic with stories', () => {
       const epic = createMockEpic();
       expect(epic.slug).toBeDefined();
       expect(epic.title).toBeDefined();
@@ -43,7 +54,7 @@ test.describe('API Mocking Infrastructure', () => {
       expect(epic.storyCounts).toBeDefined();
     });
 
-    test('createMockStoryDetail creates valid StoryDetail', async () => {
+    test('createMockStoryDetail creates valid StoryDetail', () => {
       const story = createMockStoryDetail();
       expect(story.slug).toBeDefined();
       expect(story.title).toBeDefined();
@@ -65,7 +76,9 @@ test.describe('API Mocking Infrastructure', () => {
 
       // Navigate and verify the API was mocked
       await page.goto('/');
-      await page.waitForLoadState('networkidle');
+      await expect(page.getByTestId('epic-card-skeleton')).toHaveCount(0, {
+        timeout: LOADING_TIMEOUT_MS,
+      });
 
       // The page should render our mocked epics
       await expect(page.getByText('Epic One')).toBeVisible();
@@ -79,11 +92,15 @@ test.describe('API Mocking Infrastructure', () => {
       });
 
       // Mock both the list (for initial load) and detail
-      await mockEpicList(page, [createMockEpicSummary({ slug: 'test-epic', title: 'Test Epic Detail' })]);
+      await mockEpicList(page, [
+        createMockEpicSummary({ slug: 'test-epic', title: 'Test Epic Detail' }),
+      ]);
       await mockEpicDetail(page, epic);
 
       await page.goto('/epic/test-epic');
-      await page.waitForLoadState('networkidle');
+      await expect(page.getByTestId('epic-header-skeleton')).toHaveCount(0, {
+        timeout: LOADING_TIMEOUT_MS,
+      });
 
       // The page should show the epic detail
       await expect(page.getByRole('heading', { name: 'Test Epic Detail' })).toBeVisible();
@@ -100,14 +117,16 @@ test.describe('API Mocking Infrastructure', () => {
       await mockStoryDetail(page, story);
 
       await page.goto('/epic/test-epic/story/test-story');
-      await page.waitForLoadState('networkidle');
+      await expect(page.getByTestId('story-header-skeleton')).toHaveCount(0, {
+        timeout: LOADING_TIMEOUT_MS,
+      });
 
       // The page should show the story detail
       await expect(page.getByRole('heading', { name: 'Test Story Detail' })).toBeVisible();
     });
 
     test('mockApiError mocks error responses', async ({ page }) => {
-      await mockApiError(page, '**/api/epics', 500, 'Internal Server Error');
+      await mockApiError(page, '**/api/epics', HTTP_INTERNAL_SERVER_ERROR, 'Internal Server Error');
 
       await page.goto('/');
       await page.waitForLoadState('networkidle');
@@ -121,8 +140,8 @@ test.describe('API Mocking Infrastructure', () => {
       const epics = [createMockEpicSummary({ slug: 'delayed-epic' })];
 
       // Add a 500ms delay to the response
-      await mockApiDelay(page, '**/api/epics', 500, {
-        status: 200,
+      await mockApiDelay(page, '**/api/epics', DELAY_MS, {
+        status: HTTP_OK,
         contentType: 'application/json',
         body: JSON.stringify(epics),
       });
@@ -133,7 +152,7 @@ test.describe('API Mocking Infrastructure', () => {
       const elapsed = Date.now() - start;
 
       // The request should have been delayed by at least 500ms
-      expect(elapsed).toBeGreaterThanOrEqual(500);
+      expect(elapsed).toBeGreaterThanOrEqual(DELAY_MS);
     });
   });
 });
