@@ -1,13 +1,14 @@
-import { test, expect } from '@playwright/test';
+import { expect } from '@playwright/test';
+import { test } from '../fixtures.ts';
 import {
-  createMockEpicSummary,
   createMockEpic,
+  createMockEpicSummary,
   createMockStoryDetail,
   createMockTask,
-  mockEpicList,
   mockEpicDetail,
+  mockEpicList,
   mockStoryDetail,
-} from '../utils/mock-api';
+} from '../utils/mock-api.ts';
 
 /**
  * Navigation tests for the dashboard.
@@ -15,7 +16,9 @@ import {
  */
 test.describe('Navigation', () => {
   test.describe('Epic List to Epic Detail', () => {
-    test('should navigate from epic list to epic detail when clicking an epic card', async ({ page }) => {
+    test('should navigate from epic list to epic detail when clicking an epic card', async ({
+      page,
+    }) => {
       // Setup mock data
       const epic = createMockEpicSummary({
         slug: 'test-epic',
@@ -26,7 +29,12 @@ test.describe('Navigation', () => {
         slug: 'test-epic',
         title: 'Test Epic',
         stories: [
-          createMockStoryDetail({ slug: 'story-1', title: 'Story One', status: 'ready', epicSlug: 'test-epic' }),
+          createMockStoryDetail({
+            slug: 'story-1',
+            title: 'Story One',
+            status: 'ready',
+            epicSlug: 'test-epic',
+          }),
         ],
       });
 
@@ -35,6 +43,7 @@ test.describe('Navigation', () => {
 
       // Navigate to epic list
       await page.goto('/');
+      await expect(page.getByTestId('epic-card-skeleton')).toHaveCount(0, { timeout: 10_000 });
       await expect(page.getByRole('heading', { name: 'Epics' })).toBeVisible();
       await expect(page.getByText('Test Epic')).toBeVisible();
 
@@ -63,7 +72,9 @@ test.describe('Navigation', () => {
       await mockEpicDetail(page, epicDetail);
 
       await page.goto('/');
+      await expect(page.getByTestId('epic-card-skeleton')).toHaveCount(0, { timeout: 10_000 });
       await page.getByText('Progress Epic').click();
+      await expect(page.getByTestId('epic-header-skeleton')).toHaveCount(0, { timeout: 10_000 });
 
       // Verify progress information is displayed
       await expect(page.getByText('3/6 stories completed')).toBeVisible();
@@ -77,9 +88,13 @@ test.describe('Navigation', () => {
       ];
 
       await mockEpicList(page, epics);
-      await mockEpicDetail(page, createMockEpic({ slug: 'epic-beta', title: 'Epic Beta', stories: [] }));
+      await mockEpicDetail(
+        page,
+        createMockEpic({ slug: 'epic-beta', title: 'Epic Beta', stories: [] }),
+      );
 
       await page.goto('/');
+      await expect(page.getByTestId('epic-card-skeleton')).toHaveCount(0, { timeout: 10_000 });
 
       // Click on the middle epic
       await page.getByText('Epic Beta').click();
@@ -90,7 +105,9 @@ test.describe('Navigation', () => {
   });
 
   test.describe('Epic Detail to Story Detail', () => {
-    test('should navigate from epic detail to story detail when clicking a story card', async ({ page }) => {
+    test('should navigate from epic detail to story detail when clicking a story card', async ({
+      page,
+    }) => {
       const epicDetail = createMockEpic({
         slug: 'test-epic',
         title: 'Test Epic',
@@ -117,6 +134,7 @@ test.describe('Navigation', () => {
 
       // Navigate directly to epic detail
       await page.goto('/epic/test-epic');
+      await expect(page.getByTestId('epic-header-skeleton')).toHaveCount(0, { timeout: 10_000 });
       await expect(page.getByText('My Story')).toBeVisible();
 
       // Click on the story card
@@ -151,7 +169,9 @@ test.describe('Navigation', () => {
       await mockStoryDetail(page, storyDetail);
 
       await page.goto('/epic/epic-1');
+      await expect(page.getByTestId('epic-header-skeleton')).toHaveCount(0, { timeout: 10_000 });
       await page.getByText('Blocked Story').click();
+      await expect(page.getByTestId('story-header-skeleton')).toHaveCount(0, { timeout: 10_000 });
 
       // Verify the status badge shows "Blocked"
       await expect(page.getByText('Blocked', { exact: true })).toBeVisible();
@@ -169,6 +189,7 @@ test.describe('Navigation', () => {
       await mockEpicDetail(page, epicDetail);
 
       await page.goto('/epic/my-epic');
+      await expect(page.getByTestId('epic-header-skeleton')).toHaveCount(0, { timeout: 10_000 });
 
       // Verify breadcrumb shows "Epics" link and current epic
       const breadcrumb = page.locator('nav[aria-label="Breadcrumb"]');
@@ -185,6 +206,7 @@ test.describe('Navigation', () => {
       await mockEpicDetail(page, epicDetail);
 
       await page.goto('/epic/test-epic');
+      await expect(page.getByTestId('epic-header-skeleton')).toHaveCount(0, { timeout: 10_000 });
 
       // Click on "Epics" in breadcrumb
       const breadcrumb = page.locator('nav[aria-label="Breadcrumb"]');
@@ -219,6 +241,7 @@ test.describe('Navigation', () => {
       await mockStoryDetail(page, storyDetail);
 
       await page.goto('/epic/parent-epic/story/child-story');
+      await expect(page.getByTestId('story-header-skeleton')).toHaveCount(0, { timeout: 10_000 });
 
       // Verify breadcrumb shows full path
       const breadcrumb = page.locator('nav[aria-label="Breadcrumb"]');
@@ -252,6 +275,7 @@ test.describe('Navigation', () => {
       await mockStoryDetail(page, storyDetail);
 
       await page.goto('/epic/nav-epic/story/nav-story');
+      await expect(page.getByTestId('story-header-skeleton')).toHaveCount(0, { timeout: 10_000 });
 
       // Click on the epic name in breadcrumb
       const breadcrumb = page.locator('nav[aria-label="Breadcrumb"]');
@@ -263,7 +287,9 @@ test.describe('Navigation', () => {
   });
 
   test.describe('Full Navigation Flow', () => {
-    test('should complete full navigation flow: list -> epic -> story -> back to epic -> back to list', async ({ page }) => {
+    test('should complete full navigation flow: list -> epic -> story -> back to epic -> back to list', async ({
+      page,
+    }) => {
       const epics = [
         createMockEpicSummary({
           slug: 'flow-epic',
@@ -297,15 +323,18 @@ test.describe('Navigation', () => {
 
       // Start at epic list
       await page.goto('/');
+      await expect(page.getByTestId('epic-card-skeleton')).toHaveCount(0, { timeout: 10_000 });
       await expect(page.getByRole('heading', { name: 'Epics' })).toBeVisible();
 
       // Navigate to epic detail
       await page.getByText('Flow Epic').click();
+      await expect(page.getByTestId('epic-header-skeleton')).toHaveCount(0, { timeout: 10_000 });
       await expect(page).toHaveURL('/epic/flow-epic');
       await expect(page.getByRole('heading', { name: 'Flow Epic' })).toBeVisible();
 
       // Navigate to story detail
       await page.getByText('Flow Story').click();
+      await expect(page.getByTestId('story-header-skeleton')).toHaveCount(0, { timeout: 10_000 });
       await expect(page).toHaveURL('/epic/flow-epic/story/flow-story');
       await expect(page.getByRole('heading', { name: 'Flow Story' })).toBeVisible();
       await expect(page.getByText('Flow Task')).toBeVisible();
@@ -348,8 +377,11 @@ test.describe('Navigation', () => {
 
       // Navigate forward through app
       await page.goto('/');
+      await expect(page.getByTestId('epic-card-skeleton')).toHaveCount(0, { timeout: 10_000 });
       await page.getByText('Nav Epic').click();
+      await expect(page.getByTestId('epic-header-skeleton')).toHaveCount(0, { timeout: 10_000 });
       await page.getByText('Nav Story').click();
+      await expect(page.getByTestId('story-header-skeleton')).toHaveCount(0, { timeout: 10_000 });
       await expect(page).toHaveURL('/epic/nav-epic/story/nav-story');
 
       // Go back to epic
@@ -393,10 +425,11 @@ test.describe('Navigation', () => {
       await mockStoryDetail(page, storyDetail);
 
       await page.goto('/epic/link-epic/story/link-story');
+      await expect(page.getByTestId('story-header-skeleton')).toHaveCount(0, { timeout: 10_000 });
 
       // The story detail page has an inline link to the epic in the header
-      // Find and click the epic link in the story header (above the title)
-      const epicLink = page.locator('.flex.items-center.gap-2').getByText('link-epic');
+      // Find and click the epic link in the story header (in main content, not breadcrumb)
+      const epicLink = page.getByRole('main').getByRole('link', { name: 'link-epic' });
       await epicLink.click();
 
       // Verify navigation to epic detail
