@@ -1,14 +1,14 @@
-import { Page, Route } from '@playwright/test';
+import type { Page, Route } from '@playwright/test';
 import type {
-  EpicSummary,
   Epic,
-  StoryDetail,
-  StoryCounts,
-  Task,
+  EpicSummary,
   JournalEntry,
-  TaskStatus,
+  StoryCounts,
+  StoryDetail,
   StoryStatus,
-} from '../../src/types/dashboard';
+  Task,
+  TaskStatus,
+} from '../../src/types/dashboard.ts';
 
 /**
  * API mocking utilities for Playwright integration tests.
@@ -17,15 +17,20 @@ import type {
  */
 
 // ============================================================================
+// Constants
+// ============================================================================
+
+/** HTTP 200 OK status code */
+const HTTP_OK = 200;
+
+// ============================================================================
 // Mock Data Factories
 // ============================================================================
 
 /**
  * Creates a default StoryCounts object.
  */
-export function createMockStoryCounts(
-  overrides: Partial<StoryCounts> = {}
-): StoryCounts {
+function createMockStoryCounts(overrides: Partial<StoryCounts> = {}): StoryCounts {
   return {
     ready: 0,
     inProgress: 0,
@@ -39,7 +44,7 @@ export function createMockStoryCounts(
 /**
  * Creates a mock Task with sensible defaults.
  */
-export function createMockTask(overrides: Partial<Task> = {}): Task {
+function createMockTask(overrides: Partial<Task> = {}): Task {
   return {
     id: `task-${Date.now()}`,
     title: 'Mock Task',
@@ -51,9 +56,7 @@ export function createMockTask(overrides: Partial<Task> = {}): Task {
 /**
  * Creates a mock JournalEntry with sensible defaults.
  */
-export function createMockJournalEntry(
-  overrides: Partial<JournalEntry> = {}
-): JournalEntry {
+function createMockJournalEntry(overrides: Partial<JournalEntry> = {}): JournalEntry {
   return {
     type: 'session',
     title: 'Mock Session',
@@ -67,9 +70,7 @@ export function createMockJournalEntry(
  * Creates a mock EpicSummary with sensible defaults.
  * Used for the epic list view (/api/epics endpoint).
  */
-export function createMockEpicSummary(
-  overrides: Partial<EpicSummary> = {}
-): EpicSummary {
+function createMockEpicSummary(overrides: Partial<EpicSummary> = {}): EpicSummary {
   const slug = overrides.slug || `mock-epic-${Date.now()}`;
   return {
     slug,
@@ -83,9 +84,7 @@ export function createMockEpicSummary(
  * Creates a mock StoryDetail with sensible defaults.
  * Used for story detail view (/api/stories/:epicSlug/:storySlug endpoint).
  */
-export function createMockStoryDetail(
-  overrides: Partial<StoryDetail> = {}
-): StoryDetail {
+function createMockStoryDetail(overrides: Partial<StoryDetail> = {}): StoryDetail {
   const slug = overrides.slug || `mock-story-${Date.now()}`;
   return {
     slug,
@@ -102,14 +101,14 @@ export function createMockStoryDetail(
  * Creates a mock Epic with sensible defaults.
  * Used for epic detail view (/api/epics/:slug endpoint).
  */
-export function createMockEpic(overrides: Partial<Epic> = {}): Epic {
+function createMockEpic(overrides: Partial<Epic> = {}): Epic {
   const slug = overrides.slug || `mock-epic-${Date.now()}`;
   const stories = overrides.stories || [];
 
   // Calculate story counts from stories if not provided
   const storyCounts = overrides.storyCounts || {
     ready: stories.filter((s) => s.status === 'ready').length,
-    inProgress: stories.filter((s) => s.status === 'in_progress').length,
+    inProgress: stories.filter((s) => s.status === 'inProgress').length,
     blocked: stories.filter((s) => s.status === 'blocked').length,
     completed: stories.filter((s) => s.status === 'completed').length,
     total: stories.length,
@@ -135,15 +134,15 @@ export function createMockEpic(overrides: Partial<Epic> = {}): Epic {
 // - General-purpose mocks (mockApiError, mockApiDelay, mockNetworkFailure)
 //   accept glob patterns from the caller for flexibility.
 
+/** Route pattern type - accepts glob strings, regex, or function matchers */
+type RoutePattern = string | RegExp | ((url: URL) => boolean);
+
 /**
  * Mocks the GET /api/epics endpoint to return a list of epic summaries.
  * Uses a function matcher to match exactly /api/epics (with optional trailing slash)
  * but NOT /api/epics/some-slug.
  */
-export async function mockEpicList(
-  page: Page,
-  epics: EpicSummary[]
-): Promise<void> {
+async function mockEpicList(page: Page, epics: EpicSummary[]): Promise<void> {
   await page.route(
     (url) => url.pathname === '/api/epics' || url.pathname === '/api/epics/',
     async (route: Route) => {
@@ -152,7 +151,7 @@ export async function mockEpicList(
         contentType: 'application/json',
         body: JSON.stringify(epics),
       });
-    }
+    },
   );
 }
 
@@ -160,7 +159,7 @@ export async function mockEpicList(
  * Mocks the GET /api/epics/:slug endpoint to return epic details.
  * Uses a function matcher to handle the URL properly.
  */
-export async function mockEpicDetail(page: Page, epic: Epic): Promise<void> {
+async function mockEpicDetail(page: Page, epic: Epic): Promise<void> {
   await page.route(
     (url) => url.pathname === `/api/epics/${epic.slug}`,
     async (route: Route) => {
@@ -169,7 +168,7 @@ export async function mockEpicDetail(page: Page, epic: Epic): Promise<void> {
         contentType: 'application/json',
         body: JSON.stringify(epic),
       });
-    }
+    },
   );
 }
 
@@ -177,10 +176,7 @@ export async function mockEpicDetail(page: Page, epic: Epic): Promise<void> {
  * Mocks the GET /api/stories/:epicSlug/:storySlug endpoint to return story details.
  * Uses a function matcher to handle the URL properly.
  */
-export async function mockStoryDetail(
-  page: Page,
-  story: StoryDetail
-): Promise<void> {
+async function mockStoryDetail(page: Page, story: StoryDetail): Promise<void> {
   await page.route(
     (url) => url.pathname === `/api/stories/${story.epicSlug}/${story.slug}`,
     async (route: Route) => {
@@ -189,12 +185,47 @@ export async function mockStoryDetail(
         contentType: 'application/json',
         body: JSON.stringify(story),
       });
-    }
+    },
   );
 }
 
-/** Route pattern type - accepts glob strings, regex, or function matchers */
-type RoutePattern = string | RegExp | ((url: URL) => boolean);
+/**
+ * Session data structure for mocking.
+ */
+interface MockSession {
+  name: string;
+  epicSlug: string;
+  storySlug: string;
+  status: 'running' | 'completed';
+  startTime: string;
+  outputPreview?: string;
+  outputAvailable?: boolean;
+}
+
+/**
+ * Mocks the GET /api/sessions endpoint to return session data.
+ * Handles query parameters like ?status=running.
+ */
+async function mockSessions(page: Page, sessions: MockSession[] = []): Promise<void> {
+  await page.route(
+    (url) => url.pathname === '/api/sessions' || url.pathname === '/api/sessions/',
+    async (route: Route) => {
+      const url = new URL(route.request().url());
+      const statusFilter = url.searchParams.get('status');
+
+      let filteredSessions = sessions;
+      if (statusFilter) {
+        filteredSessions = sessions.filter((s) => s.status === statusFilter);
+      }
+
+      await route.fulfill({
+        status: 200,
+        contentType: 'application/json',
+        body: JSON.stringify(filteredSessions),
+      });
+    },
+  );
+}
 
 /**
  * Mocks an API route to return an error response.
@@ -204,11 +235,11 @@ type RoutePattern = string | RegExp | ((url: URL) => boolean);
  * @param status - HTTP status code (e.g., 500, 404)
  * @param message - Error message to include in response body
  */
-export async function mockApiError(
+async function mockApiError(
   page: Page,
   routePattern: RoutePattern,
   status: number,
-  message: string
+  message: string,
 ): Promise<void> {
   await page.route(routePattern, async (route: Route) => {
     await route.fulfill({
@@ -222,7 +253,7 @@ export async function mockApiError(
 /**
  * Options for fulfilling a route response.
  */
-export interface FulfillOptions {
+interface FulfillOptions {
   status?: number;
   contentType?: string;
   body?: string;
@@ -237,16 +268,16 @@ export interface FulfillOptions {
  * @param delayMs - Delay in milliseconds before responding
  * @param fulfillOptions - Options for the response (status, contentType, body)
  */
-export async function mockApiDelay(
+async function mockApiDelay(
   page: Page,
   routePattern: RoutePattern,
   delayMs: number,
-  fulfillOptions: FulfillOptions = {}
+  fulfillOptions: FulfillOptions = {},
 ): Promise<void> {
   await page.route(routePattern, async (route: Route) => {
     await new Promise((resolve) => setTimeout(resolve, delayMs));
     await route.fulfill({
-      status: fulfillOptions.status ?? 200,
+      status: fulfillOptions.status ?? HTTP_OK,
       contentType: fulfillOptions.contentType ?? 'application/json',
       body: fulfillOptions.body ?? '[]',
     });
@@ -260,29 +291,21 @@ export async function mockApiDelay(
  * @param page - Playwright page object
  * @param routePattern - URL pattern to match (glob like `**\/api/epics`, regex, or function)
  */
-export async function mockNetworkFailure(
-  page: Page,
-  routePattern: RoutePattern
-): Promise<void> {
+async function mockNetworkFailure(page: Page, routePattern: RoutePattern): Promise<void> {
   await page.route(routePattern, async (route: Route) => {
     await route.abort('connectionrefused');
   });
 }
-
 
 // ============================================================================
 // Convenience Helpers
 // ============================================================================
 
 /**
- * Sets up a complete mock API environment with sample data.
- * Useful for quick test setup when you need a working dashboard.
+ * Creates mock epic summaries for the dashboard.
  */
-export async function setupMockDashboard(page: Page): Promise<{
-  epics: EpicSummary[];
-  epicDetails: Record<string, Epic>;
-}> {
-  const epics = [
+function createMockEpicSummaries(): EpicSummary[] {
+  return [
     createMockEpicSummary({
       slug: 'epic-one',
       title: 'Epic One',
@@ -294,73 +317,141 @@ export async function setupMockDashboard(page: Page): Promise<{
       storyCounts: { ready: 0, inProgress: 0, blocked: 1, completed: 0, total: 1 },
     }),
   ];
+}
 
-  const epicDetails: Record<string, Epic> = {
-    'epic-one': createMockEpic({
-      slug: 'epic-one',
-      title: 'Epic One',
-      content: 'This is the first epic.',
-      stories: [
-        createMockStoryDetail({
-          slug: 'story-1',
-          title: 'Story One',
-          status: 'ready',
-          epicSlug: 'epic-one',
-        }),
-        createMockStoryDetail({
-          slug: 'story-2',
-          title: 'Story Two',
-          status: 'in_progress',
-          epicSlug: 'epic-one',
-          tasks: [
-            createMockTask({ id: 't1', title: 'Task 1', status: 'completed' }),
-            createMockTask({ id: 't2', title: 'Task 2', status: 'in_progress' }),
-          ],
-        }),
-        createMockStoryDetail({
-          slug: 'story-3',
-          title: 'Story Three',
-          status: 'completed',
-          epicSlug: 'epic-one',
-        }),
-        createMockStoryDetail({
-          slug: 'story-4',
-          title: 'Story Four',
-          status: 'completed',
-          epicSlug: 'epic-one',
-        }),
-      ],
-    }),
-    'epic-two': createMockEpic({
-      slug: 'epic-two',
-      title: 'Epic Two',
-      content: 'This is the second epic.',
-      stories: [
-        createMockStoryDetail({
-          slug: 'blocked-story',
-          title: 'Blocked Story',
-          status: 'blocked',
-          epicSlug: 'epic-two',
-          journal: [
-            createMockJournalEntry({
-              type: 'blocker',
-              title: 'Need clarification',
-              content: 'Waiting for requirements.',
-            }),
-          ],
-        }),
-      ],
-    }),
+/**
+ * Creates mock epic one with its stories.
+ */
+function createMockEpicOne(): Epic {
+  return createMockEpic({
+    slug: 'epic-one',
+    title: 'Epic One',
+    content: 'This is the first epic.',
+    stories: [
+      createMockStoryDetail({
+        slug: 'story-1',
+        title: 'Story One',
+        status: 'ready',
+        epicSlug: 'epic-one',
+      }),
+      createMockStoryDetail({
+        slug: 'story-2',
+        title: 'Story Two',
+        status: 'inProgress',
+        epicSlug: 'epic-one',
+        tasks: [
+          createMockTask({ id: 't1', title: 'Task 1', status: 'completed' }),
+          createMockTask({ id: 't2', title: 'Task 2', status: 'inProgress' }),
+        ],
+      }),
+      createMockStoryDetail({
+        slug: 'story-3',
+        title: 'Story Three',
+        status: 'completed',
+        epicSlug: 'epic-one',
+      }),
+      createMockStoryDetail({
+        slug: 'story-4',
+        title: 'Story Four',
+        status: 'completed',
+        epicSlug: 'epic-one',
+      }),
+    ],
+  });
+}
+
+/**
+ * Creates mock epic two with its stories.
+ */
+function createMockEpicTwo(): Epic {
+  return createMockEpic({
+    slug: 'epic-two',
+    title: 'Epic Two',
+    content: 'This is the second epic.',
+    stories: [
+      createMockStoryDetail({
+        slug: 'blocked-story',
+        title: 'Blocked Story',
+        status: 'blocked',
+        epicSlug: 'epic-two',
+        journal: [
+          createMockJournalEntry({
+            type: 'blocker',
+            title: 'Need clarification',
+            content: 'Waiting for requirements.',
+          }),
+        ],
+      }),
+    ],
+  });
+}
+
+/**
+ * Creates standard mock data for the dashboard.
+ */
+function createMockDashboardData(): { epics: EpicSummary[]; epicDetails: Record<string, Epic> } {
+  return {
+    epics: createMockEpicSummaries(),
+    epicDetails: {
+      'epic-one': createMockEpicOne(),
+      'epic-two': createMockEpicTwo(),
+    },
   };
+}
 
-  // Set up all API mocks
+/**
+ * Sets up all API mocks for the given epic details.
+ * Also mocks the sessions endpoint with an empty array by default
+ * to prevent proxy errors during integration tests.
+ */
+async function setupApiMocks(
+  page: Page,
+  epics: EpicSummary[],
+  epicDetails: Record<string, Epic>,
+  sessions: MockSession[] = [],
+): Promise<void> {
   await mockEpicList(page, epics);
-  for (const [, epic] of Object.entries(epicDetails)) {
-    await mockEpicDetail(page, epic);
+  await mockSessions(page, sessions);
+
+  // Collect all mock promises
+  const mockPromises: Promise<void>[] = [];
+  for (const epic of Object.values(epicDetails)) {
+    mockPromises.push(mockEpicDetail(page, epic));
     for (const story of epic.stories) {
-      await mockStoryDetail(page, story);
+      mockPromises.push(mockStoryDetail(page, story));
     }
   }
 
+  await Promise.all(mockPromises);
+}
+
+/**
+ * Sets up a complete mock API environment with sample data.
+ * Useful for quick test setup when you need a working dashboard.
+ */
+async function setupMockDashboard(page: Page): Promise<{
+  epics: EpicSummary[];
+  epicDetails: Record<string, Epic>;
+}> {
+  const { epics, epicDetails } = createMockDashboardData();
+  await setupApiMocks(page, epics, epicDetails);
   return { epics, epicDetails };
 }
+
+export {
+  createMockStoryCounts,
+  createMockTask,
+  createMockJournalEntry,
+  createMockEpicSummary,
+  createMockStoryDetail,
+  createMockEpic,
+  mockEpicList,
+  mockEpicDetail,
+  mockStoryDetail,
+  mockSessions,
+  mockApiError,
+  mockApiDelay,
+  mockNetworkFailure,
+  setupMockDashboard,
+};
+export type { FulfillOptions, MockSession };
