@@ -1,29 +1,38 @@
-import { defineConfig } from 'vite';
+import path from 'node:path';
+import tailwindcss from '@tailwindcss/vite';
 import react from '@vitejs/plugin-react';
-import tailwindcss from 'tailwindcss';
-import autoprefixer from 'autoprefixer';
-import path from 'path';
+import { defineConfig } from 'vite';
 
 // https://vite.dev/config/
-export default defineConfig({
-  root: __dirname,
-  plugins: [react()],
+const config = defineConfig({
+  root: import.meta.dirname,
+  plugins: [tailwindcss(), react()],
   resolve: {
     alias: {
       '@': path.resolve(__dirname, './src'),
     },
-  },
-  css: {
-    postcss: {
-      plugins: [
-        tailwindcss(path.resolve(__dirname, 'tailwind.config.js')),
-        autoprefixer(),
-      ],
-    },
+    // Prevent multiple copies of React when dependencies (e.g., Storybook) have their own React versions
+    dedupe: ['react', 'react-dom'],
   },
   build: {
     outDir: '../../dist/client',
     emptyOutDir: true,
+    rollupOptions: {
+      output: {
+        manualChunks: {
+          'vendor-react': ['react', 'react-dom', 'react-router'],
+          'vendor-markdown': ['react-markdown', 'remark-gfm'],
+          'vendor-xstate': ['xstate', '@xstate/react'],
+          'vendor-radix': [
+            '@radix-ui/react-collapsible',
+            '@radix-ui/react-progress',
+            '@radix-ui/react-slot',
+            '@radix-ui/react-tabs',
+            '@radix-ui/react-toast',
+          ],
+        },
+      },
+    },
   },
   server: {
     proxy: {
@@ -38,3 +47,6 @@ export default defineConfig({
     },
   },
 });
+
+// Vite requires default export for config
+export { config as default };
