@@ -1,8 +1,22 @@
-import { describe, it, expect, beforeEach, afterEach } from 'vitest';
-import { spawn, execSync, type ChildProcess } from 'node:child_process';
-import { mkdtempSync, rmSync, mkdirSync, realpathSync } from 'node:fs';
+import { type ChildProcess, execSync, spawn } from 'node:child_process';
+import { mkdirSync, mkdtempSync, realpathSync, rmSync } from 'node:fs';
 import { tmpdir } from 'node:os';
 import { join } from 'node:path';
+import process from 'node:process';
+import { afterEach, beforeEach, describe, expect, it } from 'vitest';
+
+// ============================================================================
+// Test Constants
+// ============================================================================
+
+/** Timeout for CLI operations in milliseconds (5 seconds) */
+const CLI_TIMEOUT_MS = 5000;
+
+/** Base port number for random port generation */
+const TEST_PORT_BASE = 30_000;
+
+/** Range for random port offset (0-19999) */
+const TEST_PORT_RANGE = 20_000;
 
 describe('saga dashboard command', () => {
   let tempDir: string;
@@ -21,20 +35,24 @@ describe('saga dashboard command', () => {
   /**
    * Run the CLI and wait for it to exit (for error cases)
    */
-  function runCliSync(args: string, options: { cwd?: string } = {}): { stdout: string; stderr: string; exitCode: number } {
+  function runCliSync(
+    args: string,
+    options: { cwd?: string } = {},
+  ): { stdout: string; stderr: string; exitCode: number } {
     try {
       const stdout = execSync(`node ${cliPath} ${args}`, {
         cwd: options.cwd || tempDir,
         encoding: 'utf-8',
         env: { ...process.env },
-        timeout: 5000,
+        timeout: CLI_TIMEOUT_MS,
       });
       return { stdout, stderr: '', exitCode: 0 };
-    } catch (error: any) {
+    } catch (error) {
+      const spawnError = error as { stdout?: Buffer; stderr?: Buffer; status?: number };
       return {
-        stdout: error.stdout?.toString() || '',
-        stderr: error.stderr?.toString() || '',
-        exitCode: error.status || 1,
+        stdout: spawnError.stdout?.toString() || '',
+        stderr: spawnError.stderr?.toString() || '',
+        exitCode: spawnError.status || 1,
       };
     }
   }
@@ -43,7 +61,10 @@ describe('saga dashboard command', () => {
    * Start the CLI as a child process and capture output until server starts
    * @param waitForProject - If true, waits for both server startup and "Project:" line
    */
-  function startCliAsync(args: string, options: { cwd?: string; waitForProject?: boolean } = {}): Promise<{ proc: ChildProcess; stdout: string }> {
+  function startCliAsync(
+    args: string,
+    options: { cwd?: string; waitForProject?: boolean } = {},
+  ): Promise<{ proc: ChildProcess; stdout: string }> {
     return new Promise((resolve, reject) => {
       const proc = spawn('node', [cliPath, ...args.split(' ').filter(Boolean)], {
         cwd: options.cwd || tempDir,
@@ -54,7 +75,7 @@ describe('saga dashboard command', () => {
       let stdout = '';
       const timeout = setTimeout(() => {
         reject(new Error('Timeout waiting for server to start'));
-      }, 5000);
+      }, CLI_TIMEOUT_MS);
 
       proc.stdout?.on('data', (data) => {
         stdout += data.toString();
@@ -93,8 +114,11 @@ describe('saga dashboard command', () => {
   it('starts server and shows project path', async () => {
     createSagaProject(tempDir);
     // Use a random port to avoid conflicts
-    const port = 30000 + Math.floor(Math.random() * 20000);
-    const { proc, stdout } = await startCliAsync(`dashboard --port ${port}`, { cwd: tempDir, waitForProject: true });
+    const port = TEST_PORT_BASE + Math.floor(Math.random() * TEST_PORT_RANGE);
+    const { proc, stdout } = await startCliAsync(`dashboard --port ${port}`, {
+      cwd: tempDir,
+      waitForProject: true,
+    });
 
     try {
       expect(stdout).toContain(`SAGA Dashboard server running on http://localhost:${port}`);
@@ -108,7 +132,7 @@ describe('saga dashboard command', () => {
   it('uses default port 3847 when not specified', async () => {
     createSagaProject(tempDir);
     // Use a random port to avoid conflicts with the default port
-    const port = 30000 + Math.floor(Math.random() * 20000);
+    const port = TEST_PORT_BASE + Math.floor(Math.random() * TEST_PORT_RANGE);
     const { proc, stdout } = await startCliAsync(`dashboard --port ${port}`, { cwd: tempDir });
 
     try {
@@ -121,7 +145,7 @@ describe('saga dashboard command', () => {
 
   it('uses custom port when --port option is provided', async () => {
     createSagaProject(tempDir);
-    const port = 30000 + Math.floor(Math.random() * 20000);
+    const port = TEST_PORT_BASE + Math.floor(Math.random() * TEST_PORT_RANGE);
     const { proc, stdout } = await startCliAsync(`dashboard --port ${port}`, { cwd: tempDir });
 
     try {
@@ -133,7 +157,7 @@ describe('saga dashboard command', () => {
 
   it('uses custom port when --port option with equals sign', async () => {
     createSagaProject(tempDir);
-    const port = 30000 + Math.floor(Math.random() * 20000);
+    const port = TEST_PORT_BASE + Math.floor(Math.random() * TEST_PORT_RANGE);
     const { proc, stdout } = await startCliAsync(`dashboard --port=${port}`, { cwd: tempDir });
 
     try {
@@ -148,9 +172,12 @@ describe('saga dashboard command', () => {
     mkdirSync(projectDir, { recursive: true });
     createSagaProject(projectDir);
 
-    const port = 30000 + Math.floor(Math.random() * 20000);
+    const port = TEST_PORT_BASE + Math.floor(Math.random() * TEST_PORT_RANGE);
     // Run from temp dir with --path pointing to project
-    const { proc, stdout } = await startCliAsync(`--path ${projectDir} dashboard --port ${port}`, { cwd: tempDir, waitForProject: true });
+    const { proc, stdout } = await startCliAsync(`--path ${projectDir} dashboard --port ${port}`, {
+      cwd: tempDir,
+      waitForProject: true,
+    });
 
     try {
       expect(stdout).toContain('SAGA Dashboard server running on');
@@ -173,7 +200,7 @@ describe('saga dashboard command', () => {
     const subDir = join(tempDir, 'nested', 'deep', 'dir');
     mkdirSync(subDir, { recursive: true });
 
-    const port = 30000 + Math.floor(Math.random() * 20000);
+    const port = TEST_PORT_BASE + Math.floor(Math.random() * TEST_PORT_RANGE);
     const { proc, stdout } = await startCliAsync(`dashboard --port ${port}`, { cwd: subDir });
 
     try {
