@@ -6,8 +6,9 @@
  * and broadcasts real-time updates via WebSocket.
  */
 
-import { resolveProjectPath } from '../utils/project-discovery.js';
-import { startServer } from '../server/index.js';
+import process from 'node:process';
+import { startServer } from '../server/index.ts';
+import { resolveProjectPath } from '../utils/project-discovery.ts';
 
 /**
  * Options for the dashboard command
@@ -25,22 +26,23 @@ export async function dashboardCommand(options: DashboardOptions): Promise<void>
   let projectPath: string;
   try {
     projectPath = resolveProjectPath(options.path);
-  } catch (error: any) {
-    console.error(`Error: ${error.message}`);
+  } catch (error) {
+    console.error(error instanceof Error ? error.message : 'Failed to resolve SAGA project path');
     process.exit(1);
   }
 
   try {
     const server = await startServer({
       sagaRoot: projectPath,
-      port: options.port
+      port: options.port,
     });
 
+    // Output server info
+    console.log(`SAGA Dashboard server running on http://localhost:${server.port}`);
     console.log(`Project: ${projectPath}`);
 
     // Handle graceful shutdown
     process.on('SIGINT', async () => {
-      console.log('\nShutting down dashboard server...');
       await server.close();
       process.exit(0);
     });
@@ -49,8 +51,7 @@ export async function dashboardCommand(options: DashboardOptions): Promise<void>
       await server.close();
       process.exit(0);
     });
-  } catch (error: any) {
-    console.error(`Error starting server: ${error.message}`);
+  } catch (_error) {
     process.exit(1);
   }
 }
