@@ -2,11 +2,22 @@
  * Tests for saga init command
  */
 
-import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
-import { mkdtempSync, rmSync, mkdirSync, writeFileSync, realpathSync, readFileSync } from 'node:fs';
+import { execSync } from 'node:child_process';
+import {
+  existsSync,
+  mkdirSync,
+  mkdtempSync,
+  readFileSync,
+  realpathSync,
+  rmSync,
+  writeFileSync,
+} from 'node:fs';
 import { tmpdir } from 'node:os';
 import { join } from 'node:path';
-import { execSync } from 'node:child_process';
+import { afterEach, beforeEach, describe, expect, it } from 'vitest';
+
+// Top-level regex for reuse in tests
+const EXISTS_SKIP_REGEX = /exists.*skip/i;
 
 describe('init command', () => {
   let testDir: string;
@@ -31,11 +42,12 @@ describe('init command', () => {
         stdio: ['pipe', 'pipe', 'pipe'],
       });
       return { stdout, stderr: '', exitCode: 0 };
-    } catch (error: any) {
+    } catch (error) {
+      const spawnError = error as { stdout?: Buffer; stderr?: Buffer; status?: number };
       return {
-        stdout: error.stdout?.toString() || '',
-        stderr: error.stderr?.toString() || '',
-        exitCode: error.status || 1,
+        stdout: spawnError.stdout?.toString() || '',
+        stderr: spawnError.stderr?.toString() || '',
+        exitCode: spawnError.status || 1,
       };
     }
   }
@@ -52,9 +64,9 @@ describe('init command', () => {
 
       expect(result.exitCode).toBe(0);
       expect(result.stdout).toContain('.saga');
-      expect(require('node:fs').existsSync(epicsDir)).toBe(true);
-      expect(require('node:fs').existsSync(archiveDir)).toBe(true);
-      expect(require('node:fs').existsSync(worktreesDir)).toBe(true);
+      expect(existsSync(epicsDir)).toBe(true);
+      expect(existsSync(archiveDir)).toBe(true);
+      expect(existsSync(worktreesDir)).toBe(true);
     });
 
     it('should update .gitignore with worktrees pattern', () => {
@@ -104,7 +116,7 @@ describe('init command', () => {
       const result = runCli(['init']);
 
       expect(result.exitCode).toBe(0);
-      expect(require('node:fs').existsSync(join(testDir, '.saga', 'epics'))).toBe(true);
+      expect(existsSync(join(testDir, '.saga', 'epics'))).toBe(true);
     });
 
     it('should initialize from subdirectory when no .saga/ exists yet', () => {
@@ -116,16 +128,17 @@ describe('init command', () => {
       // (since there's no .saga to discover, it uses current dir)
       const cliPath = join(__dirname, '../../dist/cli.cjs');
       try {
-        const stdout = execSync(`node ${cliPath} init`, {
+        const _stdout = execSync(`node ${cliPath} init`, {
           cwd: subDir,
           encoding: 'utf-8',
           stdio: ['pipe', 'pipe', 'pipe'],
         });
         // It should initialize at the subDir since no parent .saga exists
-        expect(require('node:fs').existsSync(join(subDir, '.saga', 'epics'))).toBe(true);
-      } catch (error: any) {
+        expect(existsSync(join(subDir, '.saga', 'epics'))).toBe(true);
+      } catch (error) {
         // If it fails, check why
-        expect(error.status).toBe(0);
+        const execError = error as { status?: number };
+        expect(execError.status).toBe(0);
       }
     });
 
@@ -181,7 +194,7 @@ describe('init command', () => {
 
       expect(result.exitCode).toBe(0);
       // Directories should NOT exist
-      expect(require('node:fs').existsSync(join(testDir, '.saga'))).toBe(false);
+      expect(existsSync(join(testDir, '.saga'))).toBe(false);
     });
 
     it('should show what directories would be created', () => {
@@ -200,7 +213,7 @@ describe('init command', () => {
       const result = runCli(['init', '--dry-run', '--path', testDir]);
 
       expect(result.stdout).toContain('epics');
-      expect(result.stdout).toMatch(/exists.*skip/i);
+      expect(result.stdout).toMatch(EXISTS_SKIP_REGEX);
     });
 
     it('should show gitignore action', () => {
