@@ -9,14 +9,22 @@
  * It also updates .gitignore to ignore worktrees/
  */
 
+import {
+  appendFileSync,
+  existsSync,
+  mkdirSync,
+  readFileSync,
+  statSync,
+  writeFileSync,
+} from 'node:fs';
 import { join } from 'node:path';
-import { existsSync, statSync, mkdirSync, readFileSync, writeFileSync, appendFileSync } from 'node:fs';
-import { findProjectRoot } from '../utils/project-discovery.js';
+import process from 'node:process';
+import { findProjectRoot } from '../utils/project-discovery.ts';
 
 /**
  * Options for the init command
  */
-export interface InitOptions {
+interface InitOptions {
   path?: string;
   dryRun?: boolean;
 }
@@ -84,24 +92,19 @@ function runInitDryRun(targetPath: string): InitDryRunResult {
  * Print dry run results for init command
  */
 function printInitDryRunResults(result: InitDryRunResult): void {
-  console.log('Dry Run - Init Command');
-  console.log('======================\n');
-
-  console.log(`Target directory: ${result.targetPath}`);
-  console.log(`SAGA project exists: ${result.sagaExists ? 'yes' : 'no'}\n`);
-
+  console.log('Dry Run: SAGA Initialization');
+  console.log(`Target: ${result.targetPath}`);
+  console.log('');
   console.log('Directories:');
   for (const dir of result.directories) {
     const icon = dir.exists ? '-' : '+';
-    console.log(`  ${icon} ${dir.path}`);
-    console.log(`    Action: ${dir.action}`);
+    console.log(`  ${icon} ${dir.path}: ${dir.action}`);
   }
-
-  console.log('\nGitignore:');
-  console.log(`  Path: ${result.gitignore.path}`);
-  console.log(`  Action: ${result.gitignore.action}`);
-
-  console.log('\nDry run complete. No changes made.');
+  console.log('');
+  console.log('.gitignore:');
+  console.log(`  ${result.gitignore.action}`);
+  console.log('');
+  console.log('No changes made.');
 }
 
 /**
@@ -136,8 +139,6 @@ function createDirectoryStructure(projectRoot: string): void {
   mkdirSync(join(sagaDir, 'epics'), { recursive: true });
   mkdirSync(join(sagaDir, 'archive'), { recursive: true });
   mkdirSync(join(sagaDir, 'worktrees'), { recursive: true });
-
-  console.log('Created .saga/ directory structure');
 }
 
 /**
@@ -149,35 +150,33 @@ function updateGitignore(projectRoot: string): void {
   if (existsSync(gitignorePath)) {
     const content = readFileSync(gitignorePath, 'utf-8');
     if (content.includes(WORKTREES_PATTERN)) {
-      console.log('.gitignore already contains worktrees pattern');
+      // Pattern already exists in .gitignore, no action needed
     } else {
       appendFileSync(
         gitignorePath,
-        `\n# Claude Tasks - Worktrees (git worktree isolation for stories)\n${WORKTREES_PATTERN}\n`
+        `\n# Claude Tasks - Worktrees (git worktree isolation for stories)\n${WORKTREES_PATTERN}\n`,
       );
-      console.log('Updated .gitignore with worktrees pattern');
     }
   } else {
     writeFileSync(
       gitignorePath,
-      `# Claude Tasks - Worktrees (git worktree isolation for stories)\n${WORKTREES_PATTERN}\n`
+      `# Claude Tasks - Worktrees (git worktree isolation for stories)\n${WORKTREES_PATTERN}\n`,
     );
-    console.log('Created .gitignore with worktrees pattern');
   }
 }
 
 /**
  * Execute the init command
  */
-export async function initCommand(options: InitOptions): Promise<void> {
+function initCommand(options: InitOptions): void {
   // Validate explicit path exists and is a directory
   if (options.path) {
     if (!existsSync(options.path)) {
-      console.error(`Error: Specified path '${options.path}' does not exist`);
+      console.error(`Error: Path does not exist: ${options.path}`);
       process.exit(1);
     }
     if (!statSync(options.path).isDirectory()) {
-      console.error(`Error: Specified path '${options.path}' is not a directory`);
+      console.error(`Error: Path is not a directory: ${options.path}`);
       process.exit(1);
     }
   }
@@ -197,5 +196,10 @@ export async function initCommand(options: InitOptions): Promise<void> {
   // Update .gitignore
   updateGitignore(targetPath);
 
-  console.log(`Initialized .saga/ at ${targetPath}`);
+  // Report success
+  console.log(`Created .saga/ at ${targetPath}`);
 }
+
+// Export types and functions at the end of the file
+export type { InitOptions };
+export { initCommand };
