@@ -1,5 +1,6 @@
-import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
-import { resolve, relative } from 'node:path';
+import { relative, resolve } from 'node:path';
+import process from 'node:process';
+import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 
 // Test the internal functions by importing the module
 // We'll test the logic directly since the command reads from stdin
@@ -32,7 +33,9 @@ describe('scope-validator', () => {
     it('should allow paths within the worktree', () => {
       expect(isWithinWorktree('/project/worktree/src/file.ts', '/project/worktree')).toBe(true);
       expect(isWithinWorktree('/project/worktree/package.json', '/project/worktree')).toBe(true);
-      expect(isWithinWorktree('/project/worktree/.saga/epics/test/story.md', '/project/worktree')).toBe(true);
+      expect(
+        isWithinWorktree('/project/worktree/.saga/epics/test/story.md', '/project/worktree'),
+      ).toBe(true);
     });
 
     it('should block paths outside the worktree', () => {
@@ -43,7 +46,9 @@ describe('scope-validator', () => {
 
     it('should block parent directory traversal', () => {
       expect(isWithinWorktree('/project/worktree/../secret.txt', '/project/worktree')).toBe(false);
-      expect(isWithinWorktree('/project/worktree/../../etc/passwd', '/project/worktree')).toBe(false);
+      expect(isWithinWorktree('/project/worktree/../../etc/passwd', '/project/worktree')).toBe(
+        false,
+      );
     });
 
     it('should allow the worktree root itself', () => {
@@ -58,7 +63,7 @@ describe('scope-validator', () => {
     });
   });
 
-  describe('getFilePathFromInput', () => {
+  describe('file path extraction from hook input', () => {
     // Mirror the function logic for testing (uses hook input structure)
     const getFilePathFromInput = (hookInput: string): string | null => {
       try {
@@ -165,29 +170,39 @@ describe('scope-validator', () => {
 
       const pathEpic = parts[epicsIdx + 1];
 
-      if (parts.length > epicsIdx + 3 && parts[epicsIdx + 2] === 'stories') {
-        const pathStory = parts[epicsIdx + 3];
+      // Path indices for story folder structure
+      const storiesFolderIndex = 2;
+      const storySlugIndex = 3;
+
+      if (
+        parts.length > epicsIdx + storySlugIndex &&
+        parts[epicsIdx + storiesFolderIndex] === 'stories'
+      ) {
+        const pathStory = parts[epicsIdx + storySlugIndex];
         return pathEpic === allowedEpic && pathStory === allowedStory;
-      } else {
-        return pathEpic === allowedEpic;
       }
+      return pathEpic === allowedEpic;
     };
 
     it('should allow access to assigned story', () => {
       expect(
-        checkStoryAccess('.saga/epics/my-epic/stories/my-story/story.md', 'my-epic', 'my-story')
+        checkStoryAccess('.saga/epics/my-epic/stories/my-story/story.md', 'my-epic', 'my-story'),
       ).toBe(true);
     });
 
     it('should block access to other stories in same epic', () => {
       expect(
-        checkStoryAccess('.saga/epics/my-epic/stories/other-story/story.md', 'my-epic', 'my-story')
+        checkStoryAccess('.saga/epics/my-epic/stories/other-story/story.md', 'my-epic', 'my-story'),
       ).toBe(false);
     });
 
     it('should block access to other epics', () => {
       expect(
-        checkStoryAccess('.saga/epics/other-epic/stories/some-story/story.md', 'my-epic', 'my-story')
+        checkStoryAccess(
+          '.saga/epics/other-epic/stories/some-story/story.md',
+          'my-epic',
+          'my-story',
+        ),
       ).toBe(false);
     });
 
