@@ -4,8 +4,15 @@
  * Tests the CLI commands: sessions list, sessions status, sessions logs, sessions kill
  */
 
-import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
-import * as sessions from '../../lib/sessions.js';
+import process from 'node:process';
+import { beforeEach, describe, expect, it, vi } from 'vitest';
+import { getSessionStatus, killSession, listSessions, streamLogs } from '../../lib/sessions.ts';
+import {
+  sessionsKillCommand,
+  sessionsListCommand,
+  sessionsLogsCommand,
+  sessionsStatusCommand,
+} from './index.ts';
 
 // Mock the sessions library module
 vi.mock('../../lib/sessions.js', () => ({
@@ -23,34 +30,42 @@ describe('sessions CLI subcommands', () => {
   describe('sessionsListCommand', () => {
     it('should call listSessions and output JSON array', async () => {
       const mockSessions = [
-        { name: 'saga-epic1-story1-1234', status: 'running' as const, outputFile: '/tmp/saga-sessions/saga-epic1-story1-1234.out' },
-        { name: 'saga-epic2-story2-5678', status: 'running' as const, outputFile: '/tmp/saga-sessions/saga-epic2-story2-5678.out' },
+        {
+          name: 'saga__epic1__story1__1234',
+          status: 'running' as const,
+          outputFile: '/tmp/saga-sessions/saga__epic1__story1__1234.out',
+        },
+        {
+          name: 'saga__epic2__story2__5678',
+          status: 'running' as const,
+          outputFile: '/tmp/saga-sessions/saga__epic2__story2__5678.out',
+        },
       ];
-      vi.mocked(sessions.listSessions).mockResolvedValue(mockSessions);
-
-      const { sessionsListCommand } = await import('./index.js');
+      vi.mocked(listSessions).mockResolvedValue(mockSessions);
 
       // Capture console.log output
-      const logSpy = vi.spyOn(console, 'log').mockImplementation(() => {});
+      const logSpy = vi.spyOn(console, 'log').mockImplementation(() => {
+        // Suppress console output in tests
+      });
 
       await sessionsListCommand();
 
-      expect(sessions.listSessions).toHaveBeenCalled();
+      expect(listSessions).toHaveBeenCalled();
       expect(logSpy).toHaveBeenCalledWith(JSON.stringify(mockSessions, null, 2));
 
       logSpy.mockRestore();
     });
 
     it('should output empty array when no sessions exist', async () => {
-      vi.mocked(sessions.listSessions).mockResolvedValue([]);
+      vi.mocked(listSessions).mockResolvedValue([]);
 
-      const { sessionsListCommand } = await import('./index.js');
-
-      const logSpy = vi.spyOn(console, 'log').mockImplementation(() => {});
+      const logSpy = vi.spyOn(console, 'log').mockImplementation(() => {
+        // Suppress console output in tests
+      });
 
       await sessionsListCommand();
 
-      expect(sessions.listSessions).toHaveBeenCalled();
+      expect(listSessions).toHaveBeenCalled();
       expect(logSpy).toHaveBeenCalledWith(JSON.stringify([], null, 2));
 
       logSpy.mockRestore();
@@ -59,30 +74,30 @@ describe('sessions CLI subcommands', () => {
 
   describe('sessionsStatusCommand', () => {
     it('should call getSessionStatus and output JSON', async () => {
-      vi.mocked(sessions.getSessionStatus).mockResolvedValue({ running: true });
+      vi.mocked(getSessionStatus).mockResolvedValue({ running: true });
 
-      const { sessionsStatusCommand } = await import('./index.js');
+      const logSpy = vi.spyOn(console, 'log').mockImplementation(() => {
+        // Suppress console output in tests
+      });
 
-      const logSpy = vi.spyOn(console, 'log').mockImplementation(() => {});
+      await sessionsStatusCommand('saga__epic__story__1234');
 
-      await sessionsStatusCommand('saga-epic-story-1234');
-
-      expect(sessions.getSessionStatus).toHaveBeenCalledWith('saga-epic-story-1234');
+      expect(getSessionStatus).toHaveBeenCalledWith('saga__epic__story__1234');
       expect(logSpy).toHaveBeenCalledWith(JSON.stringify({ running: true }, null, 2));
 
       logSpy.mockRestore();
     });
 
     it('should output running: false for non-existent session', async () => {
-      vi.mocked(sessions.getSessionStatus).mockResolvedValue({ running: false });
+      vi.mocked(getSessionStatus).mockResolvedValue({ running: false });
 
-      const { sessionsStatusCommand } = await import('./index.js');
+      const logSpy = vi.spyOn(console, 'log').mockImplementation(() => {
+        // Suppress console output in tests
+      });
 
-      const logSpy = vi.spyOn(console, 'log').mockImplementation(() => {});
+      await sessionsStatusCommand('saga__nonexistent__1234');
 
-      await sessionsStatusCommand('saga-nonexistent-1234');
-
-      expect(sessions.getSessionStatus).toHaveBeenCalledWith('saga-nonexistent-1234');
+      expect(getSessionStatus).toHaveBeenCalledWith('saga__nonexistent__1234');
       expect(logSpy).toHaveBeenCalledWith(JSON.stringify({ running: false }, null, 2));
 
       logSpy.mockRestore();
@@ -91,26 +106,26 @@ describe('sessions CLI subcommands', () => {
 
   describe('sessionsLogsCommand', () => {
     it('should call streamLogs with session name', async () => {
-      vi.mocked(sessions.streamLogs).mockResolvedValue(undefined);
+      vi.mocked(streamLogs).mockResolvedValue(undefined);
 
-      const { sessionsLogsCommand } = await import('./index.js');
+      await sessionsLogsCommand('saga__epic__story__1234');
 
-      await sessionsLogsCommand('saga-epic-story-1234');
-
-      expect(sessions.streamLogs).toHaveBeenCalledWith('saga-epic-story-1234');
+      expect(streamLogs).toHaveBeenCalledWith('saga__epic__story__1234');
     });
 
     it('should handle errors from streamLogs gracefully', async () => {
-      vi.mocked(sessions.streamLogs).mockRejectedValue(new Error('Output file not found'));
+      vi.mocked(streamLogs).mockRejectedValue(new Error('Output file not found'));
 
-      const { sessionsLogsCommand } = await import('./index.js');
-
-      const errorSpy = vi.spyOn(console, 'error').mockImplementation(() => {});
+      const errorSpy = vi.spyOn(console, 'error').mockImplementation(() => {
+        // Suppress console error output in tests
+      });
       const exitSpy = vi.spyOn(process, 'exit').mockImplementation(() => {
         throw new Error('process.exit called');
       });
 
-      await expect(sessionsLogsCommand('saga-nonexistent-1234')).rejects.toThrow('process.exit called');
+      await expect(sessionsLogsCommand('saga__nonexistent__1234')).rejects.toThrow(
+        'process.exit called',
+      );
 
       expect(errorSpy).toHaveBeenCalledWith('Error: Output file not found');
       expect(exitSpy).toHaveBeenCalledWith(1);
@@ -122,30 +137,30 @@ describe('sessions CLI subcommands', () => {
 
   describe('sessionsKillCommand', () => {
     it('should call killSession and output JSON', async () => {
-      vi.mocked(sessions.killSession).mockResolvedValue({ killed: true });
+      vi.mocked(killSession).mockResolvedValue({ killed: true });
 
-      const { sessionsKillCommand } = await import('./index.js');
+      const logSpy = vi.spyOn(console, 'log').mockImplementation(() => {
+        // Suppress console output in tests
+      });
 
-      const logSpy = vi.spyOn(console, 'log').mockImplementation(() => {});
+      await sessionsKillCommand('saga__epic__story__1234');
 
-      await sessionsKillCommand('saga-epic-story-1234');
-
-      expect(sessions.killSession).toHaveBeenCalledWith('saga-epic-story-1234');
+      expect(killSession).toHaveBeenCalledWith('saga__epic__story__1234');
       expect(logSpy).toHaveBeenCalledWith(JSON.stringify({ killed: true }, null, 2));
 
       logSpy.mockRestore();
     });
 
     it('should output killed: false for non-existent session', async () => {
-      vi.mocked(sessions.killSession).mockResolvedValue({ killed: false });
+      vi.mocked(killSession).mockResolvedValue({ killed: false });
 
-      const { sessionsKillCommand } = await import('./index.js');
+      const logSpy = vi.spyOn(console, 'log').mockImplementation(() => {
+        // Suppress console output in tests
+      });
 
-      const logSpy = vi.spyOn(console, 'log').mockImplementation(() => {});
+      await sessionsKillCommand('saga__nonexistent__1234');
 
-      await sessionsKillCommand('saga-nonexistent-1234');
-
-      expect(sessions.killSession).toHaveBeenCalledWith('saga-nonexistent-1234');
+      expect(killSession).toHaveBeenCalledWith('saga__nonexistent__1234');
       expect(logSpy).toHaveBeenCalledWith(JSON.stringify({ killed: false }, null, 2));
 
       logSpy.mockRestore();
