@@ -4,23 +4,35 @@
  * Provides functions to create, list, monitor, and kill tmux sessions
  * running SAGA workers. This module is shared between the CLI and dashboard.
  *
- * Session naming convention: saga-<epic-slug>-<story-slug>-<timestamp>
+ * Session naming convention: saga__<epic-slug>__<story-slug>__<timestamp>
  * Output files stored in: /tmp/saga-sessions/<session-name>.out
  */
 
-import { spawn, spawnSync, ChildProcess } from 'node:child_process';
+import { spawn, spawnSync } from 'node:child_process';
 import { existsSync, mkdirSync, writeFileSync } from 'node:fs';
+import { readFile, stat } from 'node:fs/promises';
 import { join } from 'node:path';
+import process from 'node:process';
+
+// Constants for magic numbers
+const SESSION_NAME_PARTS_COUNT = 4;
+const PREVIEW_LINES_COUNT = 5;
+const PREVIEW_MAX_LENGTH = 500;
+
+// Top-level regex patterns
+const SLUG_PATTERN = /^[a-z0-9-]+$/;
+// Matches session name format: saga__<epic>__<story>__<timestamp>
+const SESSION_NAME_PATTERN = /^(saga__[a-z0-9_-]+):/;
 
 /**
  * Directory where session output files are stored
  */
-export const OUTPUT_DIR = '/tmp/saga-sessions';
+const OUTPUT_DIR = '/tmp/saga-sessions';
 
 /**
  * Result from creating a session
  */
-export interface CreateSessionResult {
+interface CreateSessionResult {
   sessionName: string;
   outputFile: string;
 }
@@ -28,24 +40,189 @@ export interface CreateSessionResult {
 /**
  * Session info from listing sessions
  */
-export interface SessionInfo {
+interface SessionInfo {
   name: string;
   status: 'running' | 'not_running';
   outputFile: string;
 }
 
 /**
+ * Detailed session info with parsed epic/story slugs and output data
+ * Used by the dashboard API for richer session information
+ */
+interface DetailedSessionInfo {
+  name: string;
+  epicSlug: string;
+  storySlug: string;
+  status: 'running' | 'completed';
+  outputFile: string;
+  outputAvailable: boolean;
+  startTime: Date;
+  endTime?: Date;
+  outputPreview?: string;
+}
+
+/**
+ * Result from parsing a session name
+ */
+interface ParsedSessionName {
+  epicSlug: string;
+  storySlug: string;
+}
+
+/**
  * Result from checking session status
  */
-export interface SessionStatus {
+interface SessionStatus {
   running: boolean;
 }
 
 /**
  * Result from killing a session
  */
-export interface KillSessionResult {
+interface KillSessionResult {
   killed: boolean;
+}
+
+/**
+ * Check if tmux is available on the system
+ */
+function checkTmuxAvailable(): void {
+  const result = spawnSync('which', ['tmux'], { encoding: 'utf-8' });
+  if (result.status !== 0) {
+    throw new Error('tmux is not installed or not found in PATH');
+  }
+}
+
+/**
+ * Generate wrapper script content for a SAGA session
+ */
+function generateWrapperScript(
+  commandFilePath: string,
+  outputFile: string,
+  wrapperScriptPath: string,
+): string {
+  return `#!/bin/bash
+# Auto-generated wrapper script for SAGA session
+set -e
+
+# Mark this as an internal SAGA session (used by CLI to detect it's running inside tmux)
+export SAGA_INTERNAL_SESSION=1
+
+COMMAND_FILE="${commandFilePath}"
+OUTPUT_FILE="${outputFile}"
+SCRIPT_FILE="${wrapperScriptPath}"
+
+# Read the command from file
+COMMAND="$(cat "$COMMAND_FILE")"
+
+# Cleanup temporary files on exit
+cleanup() {
+  rm -f "$COMMAND_FILE" "$SCRIPT_FILE"
+}
+trap cleanup EXIT
+
+# Execute the command with output capture
+# script syntax differs between macOS and Linux:
+#   macOS (Darwin): script -q <file> <shell> -c <command>
+#   Linux:          script -q -c <command> <file>
+if [[ "$(uname)" == "Darwin" ]]; then
+  # -F: flush output after each write (ensures immediate visibility)
+  exec script -qF "$OUTPUT_FILE" /bin/bash -c "$COMMAND"
+else
+  exec script -q -c "$COMMAND" "$OUTPUT_FILE"
+fi
+`;
+}
+
+/**
+ * Extract file timestamps from an output file
+ * @param outputFile - Path to the output file
+ * @param status - Session status to determine if endTime should be set
+ * @returns Object with startTime and optional endTime
+ */
+async function extractFileTimestamps(
+  outputFile: string,
+  status: 'running' | 'completed',
+): Promise<{ startTime: Date; endTime?: Date }> {
+  try {
+    const stats = await stat(outputFile);
+    return {
+      startTime: stats.birthtime,
+      endTime: status === 'completed' ? stats.mtime : undefined,
+    };
+  } catch {
+    // Use fallback startTime (now)
+    return { startTime: new Date() };
+  }
+}
+
+/**
+ * Generate a preview from file content (last N lines, truncated to max length)
+ * @param outputFile - Path to the output file
+ * @returns Preview string or undefined if not available
+ */
+async function generateOutputPreview(outputFile: string): Promise<string | undefined> {
+  try {
+    const content = await readFile(outputFile, 'utf-8');
+    const lines = content.split('\n').filter((line) => line.length > 0);
+    if (lines.length === 0) {
+      return undefined;
+    }
+    const lastLines = lines.slice(-PREVIEW_LINES_COUNT);
+    let preview = lastLines.join('\n');
+    if (preview.length > PREVIEW_MAX_LENGTH) {
+      // Truncate at the last newline before max chars for cleaner display
+      const truncated = preview.slice(0, PREVIEW_MAX_LENGTH);
+      const lastNewline = truncated.lastIndexOf('\n');
+      preview = lastNewline > 0 ? truncated.slice(0, lastNewline) : truncated;
+    }
+    return preview;
+  } catch {
+    // Could not read file, leave outputPreview undefined
+    return undefined;
+  }
+}
+
+/**
+ * Validate slugs for session creation and throw if invalid
+ */
+function validateSessionSlugs(epicSlug: string, storySlug: string): void {
+  if (!validateSlug(epicSlug)) {
+    throw new Error(
+      `Invalid epic slug: '${epicSlug}'. Must contain only [a-z0-9-] and not start/end with hyphen.`,
+    );
+  }
+  if (!validateSlug(storySlug)) {
+    throw new Error(
+      `Invalid story slug: '${storySlug}'. Must contain only [a-z0-9-] and not start/end with hyphen.`,
+    );
+  }
+}
+
+/**
+ * Create session files (command and wrapper script)
+ */
+function createSessionFiles(
+  sessionName: string,
+  command: string,
+): { wrapperScriptPath: string; outputFile: string } {
+  const outputFile = join(OUTPUT_DIR, `${sessionName}.out`);
+  const commandFilePath = join(OUTPUT_DIR, `${sessionName}.cmd`);
+  const wrapperScriptPath = join(OUTPUT_DIR, `${sessionName}.sh`);
+
+  // Write the raw command to a file (no shell interpretation)
+  writeFileSync(commandFilePath, command, { mode: 0o600 });
+
+  // Create wrapper script that reads the command file safely
+  const wrapperScriptContent = generateWrapperScript(
+    commandFilePath,
+    outputFile,
+    wrapperScriptPath,
+  );
+  writeFileSync(wrapperScriptPath, wrapperScriptContent, { mode: 0o700 });
+
+  return { wrapperScriptPath, outputFile };
 }
 
 /**
@@ -60,7 +237,7 @@ export interface KillSessionResult {
 export function shellEscape(str: string): string {
   // Wrap in single quotes and escape any single quotes within
   // The pattern 'text'\''more' closes the quote, adds an escaped quote, reopens
-  return "'" + str.replace(/'/g, "'\\''") + "'";
+  return `'${str.replace(/'/g, "'\\''")}'`;
 }
 
 /**
@@ -80,13 +257,13 @@ export function shellEscapeArgs(args: string[]): string {
  * @param slug - The slug to validate
  * @returns true if valid, false otherwise
  */
-export function validateSlug(slug: string): boolean {
-  if (!slug || slug.length === 0) {
+export function validateSlug(slug: unknown): boolean {
+  if (typeof slug !== 'string' || slug.length === 0) {
     return false;
   }
 
   // Must only contain lowercase letters, numbers, and hyphens
-  if (!/^[a-z0-9-]+$/.test(slug)) {
+  if (!SLUG_PATTERN.test(slug)) {
     return false;
   }
 
@@ -99,19 +276,9 @@ export function validateSlug(slug: string): boolean {
 }
 
 /**
- * Check if tmux is available on the system
- */
-function checkTmuxAvailable(): void {
-  const result = spawnSync('which', ['tmux'], { encoding: 'utf-8' });
-  if (result.status !== 0) {
-    throw new Error('tmux is not installed or not found in PATH');
-  }
-}
-
-/**
  * Create a new detached tmux session for running a command
  *
- * Creates a tmux session named: saga-<epic>-<story>-<timestamp>
+ * Creates a tmux session named: saga__<epic>__<story>__<timestamp>
  * Output is captured to: /tmp/saga-sessions/<session-name>.out
  *
  * @param epicSlug - The epic slug (validated)
@@ -119,20 +286,12 @@ function checkTmuxAvailable(): void {
  * @param command - The command to execute in the session
  * @returns Session name and output file path
  */
-export async function createSession(
+export function createSession(
   epicSlug: string,
   storySlug: string,
-  command: string
-): Promise<CreateSessionResult> {
-  // Validate slugs
-  if (!validateSlug(epicSlug)) {
-    throw new Error(`Invalid epic slug: '${epicSlug}'. Must contain only [a-z0-9-] and not start/end with hyphen.`);
-  }
-  if (!validateSlug(storySlug)) {
-    throw new Error(`Invalid story slug: '${storySlug}'. Must contain only [a-z0-9-] and not start/end with hyphen.`);
-  }
-
-  // Check tmux is available
+  command: string,
+): CreateSessionResult {
+  validateSessionSlugs(epicSlug, storySlug);
   checkTmuxAvailable();
 
   // Create output directory
@@ -141,76 +300,33 @@ export async function createSession(
   }
 
   // Generate session name with timestamp for uniqueness
+  // Uses double-underscore format for dashboard compatibility (parseSessionName requires saga__)
   const timestamp = Date.now();
-  const sessionName = `saga-${epicSlug}-${storySlug}-${timestamp}`;
-  const outputFile = join(OUTPUT_DIR, `${sessionName}.out`);
+  const sessionName = `saga__${epicSlug}__${storySlug}__${timestamp}`;
 
-  // Write command to a separate file to avoid shell injection vulnerabilities
-  const commandFilePath = join(OUTPUT_DIR, `${sessionName}.cmd`);
-  const wrapperScriptPath = join(OUTPUT_DIR, `${sessionName}.sh`);
-
-  // Write the raw command to a file (no shell interpretation)
-  writeFileSync(commandFilePath, command, { mode: 0o600 });
-
-  // Create wrapper script that reads the command file safely
-  const wrapperScriptContent = `#!/bin/bash
-# Auto-generated wrapper script for SAGA session
-set -e
-
-# Mark this as an internal SAGA session (used by CLI to detect it's running inside tmux)
-export SAGA_INTERNAL_SESSION=1
-
-COMMAND_FILE="${commandFilePath}"
-OUTPUT_FILE="${outputFile}"
-SCRIPT_FILE="${wrapperScriptPath}"
-
-# Read the command from file
-COMMAND="\$(cat "\$COMMAND_FILE")"
-
-# Cleanup temporary files on exit
-cleanup() {
-  rm -f "\$COMMAND_FILE" "\$SCRIPT_FILE"
-}
-trap cleanup EXIT
-
-# Execute the command with output capture
-# script syntax differs between macOS and Linux:
-#   macOS (Darwin): script -q <file> <shell> -c <command>
-#   Linux:          script -q -c <command> <file>
-if [[ "\$(uname)" == "Darwin" ]]; then
-  # -F: flush output after each write (ensures immediate visibility)
-  exec script -qF "\$OUTPUT_FILE" /bin/bash -c "\$COMMAND"
-else
-  exec script -q -c "\$COMMAND" "\$OUTPUT_FILE"
-fi
-`;
-
-  writeFileSync(wrapperScriptPath, wrapperScriptContent, { mode: 0o700 });
+  // Create session files
+  const { wrapperScriptPath, outputFile } = createSessionFiles(sessionName, command);
 
   // Create detached tmux session that runs the wrapper script
-  const createResult = spawnSync('tmux', [
-    'new-session',
-    '-d',              // detached
-    '-s', sessionName, // session name
-    wrapperScriptPath, // run the wrapper script
-  ], { encoding: 'utf-8' });
+  const createResult = spawnSync(
+    'tmux',
+    ['new-session', '-d', '-s', sessionName, wrapperScriptPath],
+    { encoding: 'utf-8' },
+  );
 
   if (createResult.status !== 0) {
     throw new Error(`Failed to create tmux session: ${createResult.stderr || 'unknown error'}`);
   }
 
-  return {
-    sessionName,
-    outputFile,
-  };
+  return { sessionName, outputFile };
 }
 
 /**
- * List all SAGA tmux sessions (those with saga- prefix)
+ * List all SAGA tmux sessions (those with saga__ prefix)
  *
  * @returns Array of session info
  */
-export async function listSessions(): Promise<SessionInfo[]> {
+export function listSessions(): SessionInfo[] {
   const result = spawnSync('tmux', ['ls'], { encoding: 'utf-8' });
 
   // tmux ls returns non-zero when no sessions exist
@@ -223,10 +339,8 @@ export async function listSessions(): Promise<SessionInfo[]> {
 
   for (const line of lines) {
     // tmux ls output format: "session-name: N windows ..."
-    // Session name format: saga-<epic>-<story>-<pid>
-    // - epic/story slugs contain only [a-z0-9] with hyphens between (no leading/trailing hyphens)
-    // - pid is always numeric
-    const match = line.match(/^(saga-[a-z0-9]+(?:-[a-z0-9]+)*-\d+):/);
+    // Session name format: saga__<epic>__<story>__<timestamp>
+    const match = line.match(SESSION_NAME_PATTERN);
     if (match) {
       const name = match[1];
       sessions.push({
@@ -248,7 +362,7 @@ export async function listSessions(): Promise<SessionInfo[]> {
  * @param sessionName - The session name to check
  * @returns Object with running boolean
  */
-export async function getSessionStatus(sessionName: string): Promise<SessionStatus> {
+export function getSessionStatus(sessionName: string): SessionStatus {
   const result = spawnSync('tmux', ['has-session', '-t', sessionName], {
     encoding: 'utf-8',
   });
@@ -267,7 +381,7 @@ export async function getSessionStatus(sessionName: string): Promise<SessionStat
  * @param sessionName - The session name to stream logs from
  * @returns Promise that resolves when streaming ends
  */
-export async function streamLogs(sessionName: string): Promise<void> {
+export function streamLogs(sessionName: string): Promise<void> {
   const outputFile = join(OUTPUT_DIR, `${sessionName}.out`);
 
   if (!existsSync(outputFile)) {
@@ -293,7 +407,7 @@ export async function streamLogs(sessionName: string): Promise<void> {
       reject(new Error(`Failed to stream logs: ${err.message}`));
     });
 
-    child.on('close', (code) => {
+    child.on('close', (_code) => {
       resolve();
     });
 
@@ -312,7 +426,7 @@ export async function streamLogs(sessionName: string): Promise<void> {
  * @param sessionName - The session name to kill
  * @returns Object with killed boolean
  */
-export async function killSession(sessionName: string): Promise<KillSessionResult> {
+export function killSession(sessionName: string): KillSessionResult {
   const result = spawnSync('tmux', ['kill-session', '-t', sessionName], {
     encoding: 'utf-8',
   });
@@ -321,3 +435,90 @@ export async function killSession(sessionName: string): Promise<KillSessionResul
     killed: result.status === 0,
   };
 }
+
+/**
+ * Parse a session name to extract epic and story slugs
+ *
+ * Expected format: saga__<epic-slug>__<story-slug>__<timestamp>
+ * Uses double-underscore (`__`) as delimiter to allow hyphens within slugs.
+ *
+ * @param name - The session name to parse
+ * @returns Parsed slugs or null if not a valid SAGA session name
+ */
+export function parseSessionName(name: string): ParsedSessionName | null {
+  if (!name?.startsWith('saga__')) {
+    return null;
+  }
+
+  const parts = name.split('__');
+  // Expected format: ['saga', '<epic-slug>', '<story-slug>', '<pid>']
+  if (parts.length !== SESSION_NAME_PARTS_COUNT) {
+    return null;
+  }
+
+  const [, epicSlug, storySlug, pid] = parts;
+
+  // Validate that we have non-empty slugs and a pid
+  if (!(epicSlug && storySlug && pid)) {
+    return null;
+  }
+
+  return {
+    epicSlug,
+    storySlug,
+  };
+}
+
+/**
+ * Build detailed session info from a session name and status
+ *
+ * @param name - The session name
+ * @param status - The session status ('running' or 'completed')
+ * @returns DetailedSessionInfo or null if not a valid SAGA session
+ */
+export async function buildSessionInfo(
+  name: string,
+  status: 'running' | 'completed',
+): Promise<DetailedSessionInfo | null> {
+  const parsed = parseSessionName(name);
+  if (!parsed) {
+    return null;
+  }
+
+  const outputFile = join(OUTPUT_DIR, `${name}.out`);
+  const outputAvailable = existsSync(outputFile);
+
+  let startTime = new Date();
+  let endTime: Date | undefined;
+  let outputPreview: string | undefined;
+
+  if (outputAvailable) {
+    const timestamps = await extractFileTimestamps(outputFile, status);
+    startTime = timestamps.startTime;
+    endTime = timestamps.endTime;
+    outputPreview = await generateOutputPreview(outputFile);
+  }
+
+  return {
+    name,
+    epicSlug: parsed.epicSlug,
+    storySlug: parsed.storySlug,
+    status,
+    outputFile,
+    outputAvailable,
+    startTime,
+    endTime,
+    outputPreview,
+  };
+}
+
+export { OUTPUT_DIR };
+
+export type {
+  CreateSessionResult,
+  DetailedSessionInfo,
+  KillSessionResult,
+  ParsedSessionName,
+  SessionInfo,
+  SessionStatus,
+};
