@@ -2,17 +2,30 @@
  * Tests for the Express server foundation
  */
 
-import { describe, it, expect, beforeAll, afterAll, beforeEach, afterEach } from 'vitest';
-import { startServer, type ServerInstance, type ServerConfig } from '../index.js';
-import { mkdtempSync, rmSync, mkdirSync, writeFileSync } from 'fs';
-import { tmpdir } from 'os';
-import { join } from 'path';
+import { mkdirSync, mkdtempSync, rmSync } from 'node:fs';
+import { tmpdir } from 'node:os';
+import { join } from 'node:path';
+import { afterEach, beforeEach, describe, expect, it } from 'vitest';
+import { type ServerInstance, startServer } from '../index.ts';
+
+// ============================================================================
+// Test Constants
+// ============================================================================
+
+/** Base port for random port generation */
+const PORT_BASE = 30_000;
+
+/** Range for random port generation */
+const PORT_RANGE = 20_000;
+
+/** HTTP status code for successful response */
+const HTTP_OK = 200;
 
 /**
  * Generate a random port in the ephemeral range to avoid conflicts
  */
 function getRandomPort(): number {
-  return 30000 + Math.floor(Math.random() * 20000);
+  return PORT_BASE + Math.floor(Math.random() * PORT_RANGE);
 }
 
 describe('server', () => {
@@ -75,7 +88,7 @@ describe('server', () => {
       server = await startServer({ sagaRoot: testDir, port });
 
       const response = await fetch(`http://localhost:${port}/api/health`);
-      expect(response.status).toBe(200);
+      expect(response.status).toBe(HTTP_OK);
 
       const data = await response.json();
       expect(data).toEqual({ status: 'ok' });
@@ -101,9 +114,9 @@ describe('server', () => {
       // The health endpoint should work, indicating JSON middleware is loaded
       const response = await fetch(`http://localhost:${port}/api/health`, {
         method: 'GET',
-        headers: { 'Content-Type': 'application/json' }
+        headers: { 'Content-Type': 'application/json' },
       });
-      expect(response.status).toBe(200);
+      expect(response.status).toBe(HTTP_OK);
     });
   });
 });
