@@ -6,15 +6,142 @@
  * for tasks and journal entries.
  */
 
-import { readFile } from 'fs/promises';
-import { relative } from 'path';
+import { readFile } from 'node:fs/promises';
+import { relative } from 'node:path';
 import matter from 'gray-matter';
-import {
-  scanAllStories,
-  scanEpics,
-  type ScannedStory,
-  type ScannedEpic,
-} from '../utils/saga-scanner.js';
+import { type ScannedStory, scanAllStories, scanEpics } from '../utils/saga-scanner.ts';
+
+// Regex patterns used for parsing (defined at module level for performance)
+const STORY_MD_SUFFIX_PATTERN = /\/story\.md$/;
+const TITLE_HEADING_PATTERN = /^#\s+(.+)$/m;
+const JOURNAL_SECTION_PATTERN = /^##\s+/m;
+
+// Type definitions for internal use (snake_case from YAML files)
+type InternalStoryStatus = 'ready' | 'in_progress' | 'blocked' | 'completed';
+type InternalTaskStatus = 'pending' | 'in_progress' | 'completed';
+
+// Type definitions for API responses (camelCase for frontend)
+type ApiStoryStatus = 'ready' | 'inProgress' | 'blocked' | 'completed';
+type ApiTaskStatus = 'pending' | 'inProgress' | 'completed';
+
+/**
+ * Convert snake_case status to camelCase for API response
+ */
+function toApiStoryStatus(status: InternalStoryStatus): ApiStoryStatus {
+  return status === 'in_progress' ? 'inProgress' : status;
+}
+
+/**
+ * Convert snake_case task status to camelCase for API response
+ */
+function toApiTaskStatus(status: InternalTaskStatus): ApiTaskStatus {
+  return status === 'in_progress' ? 'inProgress' : status;
+}
+
+/**
+ * Validate and normalize story status from YAML
+ */
+function validateStatus(status: unknown): InternalStoryStatus {
+  const validStatuses = ['ready', 'in_progress', 'blocked', 'completed'];
+  if (typeof status === 'string' && validStatuses.includes(status)) {
+    return status as InternalStoryStatus;
+  }
+  return 'ready'; // default
+}
+
+/**
+ * Validate and normalize task status from YAML
+ */
+function validateTaskStatus(status: unknown): InternalTaskStatus {
+  const validStatuses = ['pending', 'in_progress', 'completed'];
+  if (typeof status === 'string' && validStatuses.includes(status)) {
+    return status as InternalTaskStatus;
+  }
+  return 'pending'; // default
+}
+
+/**
+ * Parse tasks array from frontmatter
+ */
+function parseTasks(tasks: unknown): Array<{ id: string; title: string; status: ApiTaskStatus }> {
+  if (!Array.isArray(tasks)) {
+    return [];
+  }
+
+  return tasks
+    .filter((t): t is Record<string, unknown> => typeof t === 'object' && t !== null)
+    .map((t) => ({
+      id: typeof t.id === 'string' ? t.id : 'unknown',
+      title: typeof t.title === 'string' ? t.title : 'Unknown Task',
+      status: toApiTaskStatus(validateTaskStatus(t.status)),
+    }));
+}
+
+/**
+ * Convert ScannedStory to StoryDetail with rich parsing
+ */
+async function toStoryDetail(story: ScannedStory, sagaRoot: string): Promise<StoryDetail> {
+  // Parse tasks from frontmatter using gray-matter for complex YAML
+  let tasks: Task[] = [];
+  try {
+    const content = await readFile(story.storyPath, 'utf-8');
+    const parsed = matter(content);
+    tasks = parseTasks(parsed.data.tasks);
+  } catch {
+    // Use simple frontmatter if gray-matter fails
+    tasks = parseTasks(story.frontmatter.tasks);
+  }
+
+  return {
+    slug: story.slug,
+    epicSlug: story.epicSlug,
+    title: story.title,
+    status: toApiStoryStatus(validateStatus(story.status)),
+    tasks,
+    content: story.body || undefined,
+    archived: story.archived,
+    paths: {
+      storyMd: relative(sagaRoot, story.storyPath),
+      ...(story.journalPath ? { journalMd: relative(sagaRoot, story.journalPath) } : {}),
+      ...(story.worktreePath ? { worktree: relative(sagaRoot, story.worktreePath) } : {}),
+    },
+  };
+}
+
+/**
+ * Build a single Epic object from scanned data
+ */
+async function buildEpic(
+  scannedEpic: { slug: string; title: string; content: string; epicPath: string },
+  epicStories: ScannedStory[],
+  sagaRoot: string,
+): Promise<Epic> {
+  // Convert scanned stories to StoryDetail with tasks
+  const stories = await Promise.all(epicStories.map((s) => toStoryDetail(s, sagaRoot)));
+
+  // Calculate story counts (status is now camelCase from API conversion)
+  const storyCounts: StoryCounts = {
+    total: stories.length,
+    ready: stories.filter((s) => s.status === 'ready').length,
+    inProgress: stories.filter((s) => s.status === 'inProgress').length,
+    blocked: stories.filter((s) => s.status === 'blocked').length,
+    completed: stories.filter((s) => s.status === 'completed').length,
+  };
+
+  return {
+    slug: scannedEpic.slug,
+    title: scannedEpic.title,
+    content: scannedEpic.content,
+    storyCounts,
+    stories,
+    path: relative(sagaRoot, scannedEpic.epicPath),
+  };
+}
+
+// ============================================================================
+// EXPORTED INTERFACES AND FUNCTIONS
+// All exports are declared at the end of the module per useExportsLast rule
+// ============================================================================
 
 /**
  * Story counts by status
@@ -33,7 +160,7 @@ export interface StoryCounts {
 export interface Task {
   id: string;
   title: string;
-  status: 'pending' | 'in_progress' | 'completed';
+  status: 'pending' | 'inProgress' | 'completed';
 }
 
 /**
@@ -52,9 +179,10 @@ export interface StoryDetail {
   slug: string;
   epicSlug: string;
   title: string;
-  status: 'ready' | 'in_progress' | 'blocked' | 'completed';
+  status: 'ready' | 'inProgress' | 'blocked' | 'completed';
   tasks: Task[];
   journal?: JournalEntry[];
+  content?: string;
   archived?: boolean;
   paths: {
     storyMd: string;
@@ -82,75 +210,6 @@ export interface Epic extends EpicSummary {
 }
 
 /**
- * Convert ScannedStory to StoryDetail with rich parsing
- */
-async function toStoryDetail(story: ScannedStory, sagaRoot: string): Promise<StoryDetail> {
-  // Parse tasks from frontmatter using gray-matter for complex YAML
-  let tasks: Task[] = [];
-  try {
-    const content = await readFile(story.storyPath, 'utf-8');
-    const parsed = matter(content);
-    tasks = parseTasks(parsed.data.tasks);
-  } catch {
-    // Use simple frontmatter if gray-matter fails
-    tasks = parseTasks(story.frontmatter.tasks);
-  }
-
-  return {
-    slug: story.slug,
-    epicSlug: story.epicSlug,
-    title: story.title,
-    status: validateStatus(story.status),
-    tasks,
-    archived: story.archived,
-    paths: {
-      storyMd: relative(sagaRoot, story.storyPath),
-      ...(story.journalPath ? { journalMd: relative(sagaRoot, story.journalPath) } : {}),
-      ...(story.worktreePath ? { worktree: relative(sagaRoot, story.worktreePath) } : {}),
-    },
-  };
-}
-
-/**
- * Validate and normalize story status
- */
-function validateStatus(status: unknown): 'ready' | 'in_progress' | 'blocked' | 'completed' {
-  const validStatuses = ['ready', 'in_progress', 'blocked', 'completed'];
-  if (typeof status === 'string' && validStatuses.includes(status)) {
-    return status as 'ready' | 'in_progress' | 'blocked' | 'completed';
-  }
-  return 'ready'; // default
-}
-
-/**
- * Parse tasks array from frontmatter
- */
-function parseTasks(tasks: unknown): Task[] {
-  if (!Array.isArray(tasks)) {
-    return [];
-  }
-
-  return tasks
-    .filter((t): t is Record<string, unknown> => typeof t === 'object' && t !== null)
-    .map((t) => ({
-      id: typeof t.id === 'string' ? t.id : 'unknown',
-      title: typeof t.title === 'string' ? t.title : 'Unknown Task',
-      status: validateTaskStatus(t.status),
-    }));
-}
-
-/**
- * Validate and normalize task status
- */
-function validateTaskStatus(status: unknown): 'pending' | 'in_progress' | 'completed' {
-  const validStatuses = ['pending', 'in_progress', 'completed'];
-  if (typeof status === 'string' && validStatuses.includes(status)) {
-    return status as 'pending' | 'in_progress' | 'completed';
-  }
-  return 'pending'; // default
-}
-
-/**
  * Parse a story.md file into StoryDetail
  *
  * Used by websocket for real-time updates when a specific story file changes.
@@ -160,8 +219,8 @@ function validateTaskStatus(status: unknown): 'pending' | 'in_progress' | 'compl
  * @returns StoryDetail or null if file doesn't exist
  */
 export async function parseStory(storyPath: string, epicSlug: string): Promise<StoryDetail | null> {
-  const { join } = await import('path');
-  const { stat } = await import('fs/promises');
+  const { join } = await import('node:path');
+  const { stat } = await import('node:fs/promises');
 
   let content: string;
   try {
@@ -171,14 +230,16 @@ export async function parseStory(storyPath: string, epicSlug: string): Promise<S
     return null;
   }
 
-  const storyDir = storyPath.replace(/\/story\.md$/, '');
+  const storyDir = storyPath.replace(STORY_MD_SUFFIX_PATTERN, '');
   const dirName = storyDir.split('/').pop() || 'unknown';
 
   // Try to parse frontmatter, use defaults if invalid
   let frontmatter: Record<string, unknown> = {};
+  let bodyContent = '';
   try {
     const parsed = matter(content);
     frontmatter = parsed.data as Record<string, unknown>;
+    bodyContent = parsed.content;
   } catch {
     // Invalid YAML, continue with empty frontmatter
   }
@@ -186,7 +247,7 @@ export async function parseStory(storyPath: string, epicSlug: string): Promise<S
   // Extract values with defaults
   const slug = (frontmatter.id as string) || (frontmatter.slug as string) || dirName;
   const title = (frontmatter.title as string) || dirName;
-  const status = validateStatus(frontmatter.status);
+  const status = toApiStoryStatus(validateStatus(frontmatter.status));
   const tasks = parseTasks(frontmatter.tasks);
 
   // Check for journal.md
@@ -205,6 +266,7 @@ export async function parseStory(storyPath: string, epicSlug: string): Promise<S
     title,
     status,
     tasks,
+    content: bodyContent || undefined,
     paths: {
       storyMd: storyPath,
       ...(hasJournal ? { journalMd: journalPath } : {}),
@@ -223,7 +285,7 @@ export async function parseEpic(epicPath: string): Promise<string | null> {
     const content = await readFile(epicPath, 'utf-8');
 
     // Extract title from first # heading
-    const match = content.match(/^#\s+(.+)$/m);
+    const match = content.match(TITLE_HEADING_PATTERN);
     if (match) {
       return match[1].trim();
     }
@@ -251,7 +313,7 @@ export async function parseJournal(journalPath: string): Promise<JournalEntry[]>
     const entries: JournalEntry[] = [];
 
     // Split by ## headers while preserving content
-    const sections = content.split(/^##\s+/m).slice(1); // Skip content before first ##
+    const sections = content.split(JOURNAL_SECTION_PATTERN).slice(1); // Skip content before first ##
 
     for (const section of sections) {
       const lines = section.split('\n');
@@ -314,35 +376,11 @@ export async function scanSagaDirectory(sagaRoot: string): Promise<Epic[]> {
     storiesByEpic.set(story.epicSlug, existing);
   }
 
-  // Build Epic objects with rich story details
-  const epics: Epic[] = [];
-
-  for (const scannedEpic of scannedEpics) {
+  // Build Epic objects with rich story details (using Promise.all to avoid await in loop)
+  const epicPromises = scannedEpics.map((scannedEpic) => {
     const epicStories = storiesByEpic.get(scannedEpic.slug) || [];
+    return buildEpic(scannedEpic, epicStories, sagaRoot);
+  });
 
-    // Convert scanned stories to StoryDetail with tasks
-    const stories = await Promise.all(
-      epicStories.map((s) => toStoryDetail(s, sagaRoot))
-    );
-
-    // Calculate story counts
-    const storyCounts: StoryCounts = {
-      total: stories.length,
-      ready: stories.filter((s) => s.status === 'ready').length,
-      inProgress: stories.filter((s) => s.status === 'in_progress').length,
-      blocked: stories.filter((s) => s.status === 'blocked').length,
-      completed: stories.filter((s) => s.status === 'completed').length,
-    };
-
-    epics.push({
-      slug: scannedEpic.slug,
-      title: scannedEpic.title,
-      content: scannedEpic.content,
-      storyCounts,
-      stories,
-      path: relative(sagaRoot, scannedEpic.epicPath),
-    });
-  }
-
-  return epics;
+  return Promise.all(epicPromises);
 }
