@@ -7,15 +7,20 @@
  * - GET /api/stories/:epicSlug/:storySlug - returns StoryDetail with parsed journal
  */
 
-import { Router, type Request, type Response } from 'express';
+import { join } from 'node:path';
+import { type Request, type Response, Router } from 'express';
 import {
-  scanSagaDirectory,
-  parseJournal,
-  type EpicSummary,
   type Epic,
+  type EpicSummary,
+  parseJournal,
   type StoryDetail,
-} from './parser.js';
-import { join } from 'path';
+  scanSagaDirectory,
+} from './parser.ts';
+import { createSessionApiRouter } from './session-routes.ts';
+
+// HTTP status codes
+const HTTP_NOT_FOUND = 404;
+const HTTP_INTERNAL_ERROR = 500;
 
 /**
  * Get epics by scanning the saga directory
@@ -23,7 +28,7 @@ import { join } from 'path';
  * Note: The file watcher triggers WebSocket updates on changes,
  * so fresh scanning on each request ensures data consistency.
  */
-async function getEpics(sagaRoot: string): Promise<Epic[]> {
+function getEpics(sagaRoot: string): Promise<Epic[]> {
   return scanSagaDirectory(sagaRoot);
 }
 
@@ -40,14 +45,9 @@ function toEpicSummary(epic: Epic): EpicSummary {
 }
 
 /**
- * Create the API router
- *
- * @param sagaRoot - Path to the project root with .saga/ directory
- * @returns Express router with API endpoints
+ * Register epics endpoints on router
  */
-export function createApiRouter(sagaRoot: string): Router {
-  const router = Router();
-
+function registerEpicsRoutes(router: Router, sagaRoot: string): void {
   /**
    * GET /api/epics
    * Returns list of epic summaries without full story details
@@ -57,9 +57,8 @@ export function createApiRouter(sagaRoot: string): Router {
       const epics = await getEpics(sagaRoot);
       const summaries = epics.map(toEpicSummary);
       res.json(summaries);
-    } catch (error) {
-      console.error('Error fetching epics:', error);
-      res.status(500).json({ error: 'Failed to fetch epics' });
+    } catch (_error) {
+      res.status(HTTP_INTERNAL_ERROR).json({ error: 'Failed to fetch epics' });
     }
   });
 
@@ -74,17 +73,21 @@ export function createApiRouter(sagaRoot: string): Router {
       const epic = epics.find((e) => e.slug === slug);
 
       if (!epic) {
-        res.status(404).json({ error: `Epic not found: ${slug}` });
+        res.status(HTTP_NOT_FOUND).json({ error: `Epic not found: ${slug}` });
         return;
       }
 
       res.json(epic);
-    } catch (error) {
-      console.error('Error fetching epic:', error);
-      res.status(500).json({ error: 'Failed to fetch epic' });
+    } catch (_error) {
+      res.status(HTTP_INTERNAL_ERROR).json({ error: 'Failed to fetch epic' });
     }
   });
+}
 
+/**
+ * Register stories endpoints on router
+ */
+function registerStoriesRoutes(router: Router, sagaRoot: string): void {
   /**
    * GET /api/stories/:epicSlug/:storySlug
    * Returns story detail with parsed journal
@@ -96,14 +99,14 @@ export function createApiRouter(sagaRoot: string): Router {
       const epic = epics.find((e) => e.slug === epicSlug);
 
       if (!epic) {
-        res.status(404).json({ error: `Epic not found: ${epicSlug}` });
+        res.status(HTTP_NOT_FOUND).json({ error: `Epic not found: ${epicSlug}` });
         return;
       }
 
       const story = epic.stories.find((s) => s.slug === storySlug);
 
       if (!story) {
-        res.status(404).json({ error: `Story not found: ${storySlug}` });
+        res.status(HTTP_NOT_FOUND).json({ error: `Story not found: ${storySlug}` });
         return;
       }
 
@@ -117,15 +120,31 @@ export function createApiRouter(sagaRoot: string): Router {
       }
 
       res.json(story);
-    } catch (error) {
-      console.error('Error fetching story:', error);
-      res.status(500).json({ error: 'Failed to fetch story' });
+    } catch (_error) {
+      res.status(HTTP_INTERNAL_ERROR).json({ error: 'Failed to fetch story' });
     }
   });
+}
+
+/**
+ * Create the API router
+ *
+ * @param sagaRoot - Path to the project root with .saga/ directory
+ * @returns Express router with API endpoints
+ */
+export function createApiRouter(sagaRoot: string): Router {
+  const router = Router();
+
+  // Register route handlers
+  registerEpicsRoutes(router, sagaRoot);
+  registerStoriesRoutes(router, sagaRoot);
+
+  // Session routes
+  router.use(createSessionApiRouter());
 
   // Catch-all for unknown API routes
   router.use((_req: Request, res: Response) => {
-    res.status(404).json({ error: 'API endpoint not found' });
+    res.status(HTTP_NOT_FOUND).json({ error: 'API endpoint not found' });
   });
 
   return router;
