@@ -13,11 +13,39 @@
  * - unsubscribe:story: Unsubscribe from story updates
  */
 
-import { WebSocketServer, WebSocket, type RawData } from 'ws';
-import type { Server as HttpServer } from 'http';
-import { scanSagaDirectory, parseStory, parseJournal, type StoryDetail, type EpicSummary } from './parser.js';
-import { createSagaWatcher, type SagaWatcher, type WatcherEvent } from './watcher.js';
-import { join, relative } from 'path';
+import type { Server as HttpServer } from 'node:http';
+import { join, relative } from 'node:path';
+import { type RawData, WebSocket, WebSocketServer } from 'ws';
+
+import {
+  LogStreamManager,
+  type LogsDataMessage,
+  type LogsErrorMessage,
+} from '../lib/log-stream-manager.ts';
+import {
+  type SessionsUpdatedMessage,
+  startSessionPolling,
+  stopSessionPolling,
+} from '../lib/session-polling.ts';
+import {
+  type EpicSummary,
+  parseJournal,
+  parseStory,
+  type StoryDetail,
+  scanSagaDirectory,
+} from './parser.ts';
+import { createSagaWatcher, type SagaWatcher, type WatcherEvent } from './watcher.ts';
+
+// ============================================================================
+// Constants
+// ============================================================================
+
+/** Heartbeat interval for checking client connections (30 seconds) */
+const HEARTBEAT_INTERVAL_MS = 30_000;
+
+// ============================================================================
+// Types
+// ============================================================================
 
 /**
  * Subscription key for a story
@@ -36,7 +64,7 @@ interface ClientState {
 /**
  * WebSocket server instance
  */
-export interface WebSocketInstance {
+interface WebSocketInstance {
   /** Broadcast epics:updated to all clients */
   broadcastEpicsUpdated(epics: EpicSummary[]): void;
   /** Broadcast story:updated to subscribed clients */
@@ -53,6 +81,7 @@ interface ClientMessage {
   data?: {
     epicSlug?: string;
     storySlug?: string;
+    sessionName?: string;
   };
 }
 
@@ -64,6 +93,10 @@ interface ServerMessage {
   data: unknown;
 }
 
+// ============================================================================
+// Helper Functions
+// ============================================================================
+
 /**
  * Create a story key from epic and story slugs
  */
@@ -74,7 +107,12 @@ function makeStoryKey(epicSlug: string, storySlug: string): StoryKey {
 /**
  * Convert Epic to EpicSummary (remove stories and content)
  */
-function toEpicSummary(epic: { slug: string; title: string; storyCounts: EpicSummary['storyCounts']; path: string }): EpicSummary {
+function toEpicSummary(epic: {
+  slug: string;
+  title: string;
+  storyCounts: EpicSummary['storyCounts'];
+  path: string;
+}): EpicSummary {
   return {
     slug: epic.slug,
     title: epic.title,
@@ -84,218 +122,385 @@ function toEpicSummary(epic: { slug: string; title: string; storyCounts: EpicSum
 }
 
 /**
- * Create and attach WebSocket server to HTTP server
- *
- * @param httpServer - The HTTP server to attach to
- * @param sagaRoot - Path to the project root containing .saga/ directory
- * @returns WebSocketInstance for broadcasting updates
+ * Send message to a single client
  */
-export async function createWebSocketServer(
-  httpServer: HttpServer,
-  sagaRoot: string
-): Promise<WebSocketInstance> {
-  // Create WebSocket server attached to HTTP server
-  const wss = new WebSocketServer({ server: httpServer });
+function sendToClient(ws: WebSocket, message: ServerMessage): void {
+  if (ws.readyState === WebSocket.OPEN) {
+    ws.send(JSON.stringify(message));
+  }
+}
 
-  // Track connected clients
-  const clients = new Map<WebSocket, ClientState>();
+/**
+ * Send log messages to a specific client
+ */
+function sendLogMessage(ws: WebSocket, message: LogsDataMessage | LogsErrorMessage): void {
+  if (ws.readyState === WebSocket.OPEN) {
+    ws.send(JSON.stringify({ event: message.type, data: message }));
+  }
+}
 
-  // Create file watcher
-  let watcher: SagaWatcher | null = null;
+// ============================================================================
+// Message Handlers
+// ============================================================================
 
-  try {
-    watcher = await createSagaWatcher(sagaRoot);
-  } catch (err) {
-    console.warn('Failed to create file watcher:', err);
-    // Continue without file watching
+/**
+ * Handle story subscription message
+ */
+function handleStorySubscription(
+  state: ClientState,
+  data: ClientMessage['data'],
+  subscribe: boolean,
+): void {
+  const { epicSlug, storySlug } = data || {};
+  if (epicSlug && storySlug) {
+    const key = makeStoryKey(epicSlug, storySlug);
+    if (subscribe) {
+      state.subscribedStories.add(key);
+    } else {
+      state.subscribedStories.delete(key);
+    }
+  }
+}
+
+/**
+ * Handle logs subscription message
+ */
+function handleLogsSubscription(
+  logStreamManager: LogStreamManager,
+  ws: WebSocket,
+  data: ClientMessage['data'],
+  subscribe: boolean,
+): void {
+  const { sessionName } = data || {};
+  if (sessionName) {
+    if (subscribe) {
+      logStreamManager.subscribe(sessionName, ws);
+    } else {
+      logStreamManager.unsubscribe(sessionName, ws);
+    }
+  }
+}
+
+/**
+ * Process client message and dispatch to appropriate handler
+ */
+function processClientMessage(
+  message: ClientMessage,
+  state: ClientState,
+  logStreamManager: LogStreamManager,
+): void {
+  switch (message.event) {
+    case 'subscribe:story':
+      handleStorySubscription(state, message.data, true);
+      break;
+    case 'unsubscribe:story':
+      handleStorySubscription(state, message.data, false);
+      break;
+    case 'subscribe:logs':
+      handleLogsSubscription(logStreamManager, state.ws, message.data, true);
+      break;
+    case 'unsubscribe:logs':
+      handleLogsSubscription(logStreamManager, state.ws, message.data, false);
+      break;
+    default:
+      // Unknown event, ignore
+      break;
+  }
+}
+
+// ============================================================================
+// Story Change Handlers
+// ============================================================================
+
+/**
+ * Check if any client is subscribed to a story
+ */
+function hasSubscribers(clients: Map<WebSocket, ClientState>, storyKey: StoryKey): boolean {
+  for (const [, state] of clients) {
+    if (state.subscribedStories.has(storyKey)) {
+      return true;
+    }
+  }
+  return false;
+}
+
+/**
+ * Get story path based on archived status
+ */
+function getStoryPath(
+  sagaRoot: string,
+  epicSlug: string,
+  storySlug: string,
+  archived: boolean | undefined,
+): string {
+  return archived
+    ? join(sagaRoot, '.saga', 'archive', epicSlug, storySlug, 'story.md')
+    : join(sagaRoot, '.saga', 'epics', epicSlug, 'stories', storySlug, 'story.md');
+}
+
+/**
+ * Parse and enrich story data with relative paths and journal
+ */
+async function parseAndEnrichStory(
+  sagaRoot: string,
+  storyPath: string,
+  epicSlug: string,
+  archived: boolean | undefined,
+): Promise<StoryDetail | null> {
+  const story = await parseStory(storyPath, epicSlug);
+  if (!story) {
+    return null;
   }
 
-  // Heartbeat interval (30 seconds)
-  const heartbeatInterval = setInterval(() => {
+  // Make paths relative
+  story.paths.storyMd = relative(sagaRoot, story.paths.storyMd);
+  if (story.paths.journalMd) {
+    story.paths.journalMd = relative(sagaRoot, story.paths.journalMd);
+  }
+
+  story.archived = archived;
+
+  // Parse journal if it exists
+  if (story.paths.journalMd) {
+    const journalPath = join(sagaRoot, story.paths.journalMd);
+    const journal = await parseJournal(journalPath);
+    if (journal.length > 0) {
+      story.journal = journal;
+    }
+  }
+
+  return story;
+}
+
+/**
+ * Handle story change event
+ */
+async function handleStoryChangeEvent(
+  event: WatcherEvent,
+  sagaRoot: string,
+  clients: Map<WebSocket, ClientState>,
+  broadcastToSubscribers: (storyKey: StoryKey, message: ServerMessage) => void,
+  handleEpicChange: () => Promise<void>,
+): Promise<void> {
+  const { epicSlug, storySlug, archived } = event;
+  if (!storySlug) {
+    return;
+  }
+
+  const storyKey = makeStoryKey(epicSlug, storySlug);
+
+  if (!hasSubscribers(clients, storyKey)) {
+    await handleEpicChange();
+    return;
+  }
+
+  try {
+    const storyPath = getStoryPath(sagaRoot, epicSlug, storySlug, archived);
+    const story = await parseAndEnrichStory(sagaRoot, storyPath, epicSlug, archived);
+
+    if (story) {
+      broadcastToSubscribers(storyKey, { event: 'story:updated', data: story });
+    }
+
+    await handleEpicChange();
+  } catch {
+    // Silently ignore story parsing errors
+  }
+}
+
+// ============================================================================
+// Connection Handlers
+// ============================================================================
+
+/**
+ * Handle client message event
+ */
+function handleClientMessage(
+  data: RawData,
+  state: ClientState,
+  logStreamManager: LogStreamManager,
+): void {
+  try {
+    const message = JSON.parse(data.toString()) as ClientMessage & { type?: string };
+    // Handle ping/pong heartbeat (client sends { type: 'ping' })
+    if (message.type === 'ping') {
+      sendToClient(state.ws, { event: 'pong', data: null });
+      return;
+    }
+    if (message.event) {
+      processClientMessage(message, state, logStreamManager);
+    }
+  } catch {
+    // Malformed JSON message from client - ignore silently
+  }
+}
+
+/**
+ * Set up event handlers for a client connection
+ */
+function setupClientHandlers(
+  ws: WebSocket,
+  state: ClientState,
+  clients: Map<WebSocket, ClientState>,
+  logStreamManager: LogStreamManager,
+): void {
+  ws.on('pong', () => {
+    state.isAlive = true;
+  });
+
+  ws.on('message', (data: RawData) => {
+    handleClientMessage(data, state, logStreamManager);
+  });
+
+  ws.on('close', () => {
+    clients.delete(ws);
+    logStreamManager.handleClientDisconnect(ws);
+  });
+
+  ws.on('error', () => {
+    clients.delete(ws);
+    logStreamManager.handleClientDisconnect(ws);
+  });
+}
+
+/**
+ * Handle new client connection
+ */
+function handleNewConnection(
+  ws: WebSocket,
+  clients: Map<WebSocket, ClientState>,
+  logStreamManager: LogStreamManager,
+): void {
+  const state: ClientState = {
+    ws,
+    subscribedStories: new Set(),
+    isAlive: true,
+  };
+  clients.set(ws, state);
+  setupClientHandlers(ws, state, clients, logStreamManager);
+}
+
+// ============================================================================
+// Watcher Setup
+// ============================================================================
+
+/**
+ * Set up watcher event handlers
+ */
+function setupWatcherHandlers(
+  watcher: SagaWatcher,
+  sagaRoot: string,
+  clients: Map<WebSocket, ClientState>,
+  broadcast: (message: ServerMessage) => void,
+  broadcastToSubscribers: (storyKey: StoryKey, message: ServerMessage) => void,
+): void {
+  const handleEpicChange = async () => {
+    try {
+      const epics = await scanSagaDirectory(sagaRoot);
+      const summaries = epics.map(toEpicSummary);
+      broadcast({ event: 'epics:updated', data: summaries });
+    } catch {
+      // Silently ignore epic scanning errors
+    }
+  };
+
+  watcher.on('epic:added', handleEpicChange);
+  watcher.on('epic:changed', handleEpicChange);
+  watcher.on('epic:removed', handleEpicChange);
+
+  const handleStoryChange = (event: WatcherEvent) => {
+    handleStoryChangeEvent(event, sagaRoot, clients, broadcastToSubscribers, handleEpicChange);
+  };
+
+  watcher.on('story:added', handleStoryChange);
+  watcher.on('story:changed', handleStoryChange);
+  watcher.on('story:removed', handleStoryChange);
+
+  watcher.on('error', () => {
+    // Silently ignore watcher errors - file watching is best-effort
+  });
+}
+
+// ============================================================================
+// Session Polling Setup
+// ============================================================================
+
+/**
+ * Set up session polling with completion detection
+ */
+function setupSessionPolling(
+  broadcast: (message: ServerMessage) => void,
+  logStreamManager: LogStreamManager,
+): void {
+  let previousSessionStates = new Map<string, 'running' | 'completed'>();
+
+  startSessionPolling((msg: SessionsUpdatedMessage) => {
+    broadcast({ event: msg.type, data: msg.data });
+
+    const currentStates = new Map<string, 'running' | 'completed'>();
+    for (const session of msg.data) {
+      currentStates.set(session.name, session.status);
+
+      const previousStatus = previousSessionStates.get(session.name);
+      if (previousStatus === 'running' && session.status === 'completed') {
+        logStreamManager.notifySessionCompleted(session.name);
+      }
+    }
+    previousSessionStates = currentStates;
+  });
+}
+
+// ============================================================================
+// Heartbeat Setup
+// ============================================================================
+
+/**
+ * Set up heartbeat interval for checking client connections
+ */
+function setupHeartbeat(clients: Map<WebSocket, ClientState>): ReturnType<typeof setInterval> {
+  return setInterval(() => {
     for (const [ws, state] of clients) {
       if (!state.isAlive) {
-        // Connection didn't respond to last ping, terminate
         clients.delete(ws);
         ws.terminate();
         continue;
       }
-
       state.isAlive = false;
       ws.ping();
     }
-  }, 30000);
+  }, HEARTBEAT_INTERVAL_MS);
+}
 
-  // Helper to send message to client
-  function sendToClient(ws: WebSocket, message: ServerMessage): void {
-    if (ws.readyState === WebSocket.OPEN) {
-      ws.send(JSON.stringify(message));
-    }
-  }
+// ============================================================================
+// WebSocket Instance Factory
+// ============================================================================
 
-  // Helper to broadcast to all clients
-  function broadcast(message: ServerMessage): void {
-    for (const [ws] of clients) {
-      sendToClient(ws, message);
-    }
-  }
+/**
+ * Internal state for WebSocket server
+ */
+interface WebSocketServerState {
+  wss: WebSocketServer;
+  clients: Map<WebSocket, ClientState>;
+  watcher: SagaWatcher | null;
+  logStreamManager: LogStreamManager;
+  heartbeatInterval: ReturnType<typeof setInterval>;
+  broadcast: (message: ServerMessage) => void;
+  broadcastToSubscribers: (storyKey: StoryKey, message: ServerMessage) => void;
+}
 
-  // Helper to broadcast to clients subscribed to a story
-  function broadcastToSubscribers(storyKey: StoryKey, message: ServerMessage): void {
-    for (const [ws, state] of clients) {
-      if (state.subscribedStories.has(storyKey)) {
-        sendToClient(ws, message);
-      }
-    }
-  }
-
-  // Handle new connections
-  wss.on('connection', (ws: WebSocket) => {
-    // Initialize client state
-    const state: ClientState = {
-      ws,
-      subscribedStories: new Set(),
-      isAlive: true,
-    };
-    clients.set(ws, state);
-
-    // Handle pong (heartbeat response)
-    ws.on('pong', () => {
-      state.isAlive = true;
-    });
-
-    // Handle messages from client
-    ws.on('message', (data: RawData) => {
-      try {
-        const message = JSON.parse(data.toString()) as ClientMessage;
-
-        if (!message.event) {
-          return; // Ignore messages without event
-        }
-
-        switch (message.event) {
-          case 'subscribe:story': {
-            const { epicSlug, storySlug } = message.data || {};
-            if (epicSlug && storySlug) {
-              const key = makeStoryKey(epicSlug, storySlug);
-              state.subscribedStories.add(key);
-            }
-            break;
-          }
-
-          case 'unsubscribe:story': {
-            const { epicSlug, storySlug } = message.data || {};
-            if (epicSlug && storySlug) {
-              const key = makeStoryKey(epicSlug, storySlug);
-              state.subscribedStories.delete(key);
-            }
-            break;
-          }
-
-          default:
-            // Unknown event, ignore
-            break;
-        }
-      } catch {
-        // Malformed message, ignore
-      }
-    });
-
-    // Handle disconnection
-    ws.on('close', () => {
-      clients.delete(ws);
-    });
-
-    // Handle errors
-    ws.on('error', (err) => {
-      console.error('WebSocket error:', err);
-      clients.delete(ws);
-    });
-  });
-
-  // Handle watcher events
-  if (watcher) {
-    // Epic events - broadcast full epic list
-    const handleEpicChange = async () => {
-      try {
-        const epics = await scanSagaDirectory(sagaRoot);
-        const summaries = epics.map(toEpicSummary);
-        broadcast({ event: 'epics:updated', data: summaries });
-      } catch (err) {
-        console.error('Error broadcasting epic update:', err);
-      }
-    };
-
-    watcher.on('epic:added', handleEpicChange);
-    watcher.on('epic:changed', handleEpicChange);
-    watcher.on('epic:removed', handleEpicChange);
-
-    // Story events - broadcast to subscribed clients
-    const handleStoryChange = async (event: WatcherEvent) => {
-      const { epicSlug, storySlug, archived } = event;
-      if (!storySlug) return;
-
-      const storyKey = makeStoryKey(epicSlug, storySlug);
-
-      // Check if any client is subscribed to this story
-      let hasSubscribers = false;
-      for (const [, state] of clients) {
-        if (state.subscribedStories.has(storyKey)) {
-          hasSubscribers = true;
-          break;
-        }
-      }
-
-      if (!hasSubscribers) {
-        // Also trigger epics:updated for story changes (counts may change)
-        await handleEpicChange();
-        return;
-      }
-
-      try {
-        // Parse the updated story
-        const storyPath = archived
-          ? join(sagaRoot, '.saga', 'archive', epicSlug, storySlug, 'story.md')
-          : join(sagaRoot, '.saga', 'epics', epicSlug, 'stories', storySlug, 'story.md');
-
-        const story = await parseStory(storyPath, epicSlug);
-
-        if (story) {
-          // Make paths relative
-          story.paths.storyMd = relative(sagaRoot, story.paths.storyMd);
-          if (story.paths.journalMd) {
-            story.paths.journalMd = relative(sagaRoot, story.paths.journalMd);
-          }
-
-          story.archived = archived;
-
-          // Parse journal if it exists
-          if (story.paths.journalMd) {
-            const journalPath = join(sagaRoot, story.paths.journalMd);
-            const journal = await parseJournal(journalPath);
-            if (journal.length > 0) {
-              story.journal = journal;
-            }
-          }
-
-          // Broadcast to subscribed clients
-          broadcastToSubscribers(storyKey, { event: 'story:updated', data: story });
-        }
-
-        // Also trigger epics:updated (story counts may have changed)
-        await handleEpicChange();
-      } catch (err) {
-        console.error('Error broadcasting story update:', err);
-      }
-    };
-
-    watcher.on('story:added', handleStoryChange);
-    watcher.on('story:changed', handleStoryChange);
-    watcher.on('story:removed', handleStoryChange);
-
-    watcher.on('error', (err: Error) => {
-      console.error('Watcher error:', err);
-    });
-  }
+/**
+ * Create WebSocket instance with broadcast and close methods
+ */
+function createWebSocketInstance(state: WebSocketServerState): WebSocketInstance {
+  const {
+    wss,
+    clients,
+    watcher,
+    logStreamManager,
+    heartbeatInterval,
+    broadcast,
+    broadcastToSubscribers,
+  } = state;
 
   return {
     broadcastEpicsUpdated(epics: EpicSummary[]): void {
@@ -309,19 +514,18 @@ export async function createWebSocketServer(
 
     async close(): Promise<void> {
       clearInterval(heartbeatInterval);
+      stopSessionPolling();
+      await logStreamManager.dispose();
 
-      // Close all client connections
       for (const [ws] of clients) {
         ws.close();
       }
       clients.clear();
 
-      // Close watcher
       if (watcher) {
         await watcher.close();
       }
 
-      // Close WebSocket server
       return new Promise((resolve, reject) => {
         wss.close((err) => {
           if (err) {
@@ -334,3 +538,73 @@ export async function createWebSocketServer(
     },
   };
 }
+
+// ============================================================================
+// Main Factory Function
+// ============================================================================
+
+/**
+ * Create and attach WebSocket server to HTTP server
+ *
+ * @param httpServer - The HTTP server to attach to
+ * @param sagaRoot - Path to the project root containing .saga/ directory
+ * @returns WebSocketInstance for broadcasting updates
+ */
+async function createWebSocketServer(
+  httpServer: HttpServer,
+  sagaRoot: string,
+): Promise<WebSocketInstance> {
+  const wss = new WebSocketServer({ server: httpServer });
+  const clients = new Map<WebSocket, ClientState>();
+
+  let watcher: SagaWatcher | null = null;
+  try {
+    watcher = await createSagaWatcher(sagaRoot);
+  } catch {
+    // Continue without file watching
+  }
+
+  const logStreamManager = new LogStreamManager(sendLogMessage);
+
+  const broadcast = (message: ServerMessage): void => {
+    for (const [ws] of clients) {
+      sendToClient(ws, message);
+    }
+  };
+
+  const broadcastToSubscribers = (storyKey: StoryKey, message: ServerMessage): void => {
+    for (const [ws, state] of clients) {
+      if (state.subscribedStories.has(storyKey)) {
+        sendToClient(ws, message);
+      }
+    }
+  };
+
+  const heartbeatInterval = setupHeartbeat(clients);
+  setupSessionPolling(broadcast, logStreamManager);
+
+  wss.on('connection', (ws: WebSocket) => {
+    handleNewConnection(ws, clients, logStreamManager);
+  });
+
+  if (watcher) {
+    setupWatcherHandlers(watcher, sagaRoot, clients, broadcast, broadcastToSubscribers);
+  }
+
+  return createWebSocketInstance({
+    wss,
+    clients,
+    watcher,
+    logStreamManager,
+    heartbeatInterval,
+    broadcast,
+    broadcastToSubscribers,
+  });
+}
+
+// ============================================================================
+// Exports
+// ============================================================================
+
+export { createWebSocketServer };
+export type { WebSocketInstance };
