@@ -2,19 +2,25 @@
  * Tests for finder utility
  */
 
-import { describe, it, expect, beforeEach, afterEach } from 'vitest';
-import { mkdtempSync, rmSync, mkdirSync, writeFileSync, realpathSync } from 'node:fs';
+import { mkdirSync, mkdtempSync, realpathSync, rmSync, writeFileSync } from 'node:fs';
 import { tmpdir } from 'node:os';
 import { join } from 'node:path';
-import {
-  parseFrontmatter,
-  extractContext,
-  findEpic,
-  findStory,
-  type EpicInfo,
-  type StoryInfo,
-  type FindResult,
-} from './finder.js';
+import { afterEach, beforeEach, describe, expect, it } from 'vitest';
+import { extractContext, findEpic, findStory } from './finder.ts';
+import { parseFrontmatter } from './saga-scanner.ts';
+
+// ============================================================================
+// Test Constants
+// ============================================================================
+
+/** Length for testing context truncation */
+const TEST_LONG_CONTEXT_LENGTH = 400;
+
+/** Maximum length for context extraction in tests */
+const TEST_MAX_CONTEXT_LENGTH = 300;
+
+/** Regex pattern for error message matching in findStory tests */
+const WORKTREES_EPICS_ERROR_REGEX = /worktrees|epics/;
 
 describe('parseFrontmatter', () => {
   it('should parse simple frontmatter', () => {
@@ -95,12 +101,12 @@ This is the context section with important information.
   });
 
   it('should truncate long context', () => {
-    const longContext = 'A'.repeat(400);
+    const longContext = 'A'.repeat(TEST_LONG_CONTEXT_LENGTH);
     const body = `## Context\n\n${longContext}\n\n## Tasks`;
 
-    const result = extractContext(body, 300);
+    const result = extractContext(body, TEST_MAX_CONTEXT_LENGTH);
 
-    expect(result.length).toBe(300);
+    expect(result.length).toBe(TEST_MAX_CONTEXT_LENGTH);
     expect(result.endsWith('...')).toBe(true);
   });
 
@@ -266,7 +272,7 @@ describe('findStory', () => {
     epicSlug: string,
     storySlug: string,
     frontmatter: Record<string, string>,
-    body: string = ''
+    body = '',
   ): void {
     const worktreesDir = join(testDir, '.saga', 'worktrees');
     const storyDir = join(
@@ -277,7 +283,7 @@ describe('findStory', () => {
       'epics',
       epicSlug,
       'stories',
-      storySlug
+      storySlug,
     );
     mkdirSync(storyDir, { recursive: true });
 
@@ -360,7 +366,7 @@ This story implements the login feature for the application.
 
 ## Tasks
 
-- Task 1`
+- Task 1`,
     );
 
     const result = await findStory(testDir, 'implement-login');
@@ -368,7 +374,7 @@ This story implements the login feature for the application.
     expect(result.found).toBe(true);
     if (result.found) {
       expect(result.data.context).toBe(
-        'This story implements the login feature for the application.'
+        'This story implements the login feature for the application.',
       );
     }
   });
@@ -414,7 +420,7 @@ This story implements the login feature for the application.
 
     expect(result.found).toBe(false);
     if (!result.found && 'error' in result) {
-      expect(result.error).toMatch(/worktrees|epics/);
+      expect(result.error).toMatch(WORKTREES_EPICS_ERROR_REGEX);
     }
   });
 
